itest: split harness test into assertion and rpc
