+++ lntest/harness.go
@@ -1,10 +1,8 @@
 package lntest
 
 import (
-	"bytes"
 	"context"
 	"crypto/rand"
-	"encoding/hex"
 	"fmt"
 	"math"
 	"strings"
@@ -13,26 +11,16 @@ import (
 
 	"github.com/btcsuite/btcd/btcutil"
 	"github.com/btcsuite/btcd/chaincfg/chainhash"
-	"github.com/btcsuite/btcd/txscript"
 	"github.com/btcsuite/btcd/wire"
 	"github.com/go-errors/errors"
 	"github.com/lightningnetwork/lnd/channeldb"
 	"github.com/lightningnetwork/lnd/kvdb/etcd"
 	"github.com/lightningnetwork/lnd/lnrpc"
-	"github.com/lightningnetwork/lnd/lnrpc/chainrpc"
-	"github.com/lightningnetwork/lnd/lnrpc/invoicesrpc"
-	"github.com/lightningnetwork/lnd/lnrpc/peersrpc"
 	"github.com/lightningnetwork/lnd/lnrpc/routerrpc"
-	"github.com/lightningnetwork/lnd/lnrpc/signrpc"
-	"github.com/lightningnetwork/lnd/lnrpc/walletrpc"
-	"github.com/lightningnetwork/lnd/lnrpc/watchtowerrpc"
-	"github.com/lightningnetwork/lnd/lnrpc/wtclientrpc"
 	"github.com/lightningnetwork/lnd/lntest/wait"
 	"github.com/lightningnetwork/lnd/lntypes"
 	"github.com/lightningnetwork/lnd/lnwallet/chainfee"
-	"github.com/lightningnetwork/lnd/lnwire"
 	"github.com/stretchr/testify/require"
-	"google.golang.org/protobuf/proto"
 )
 
 const (
@@ -263,28 +251,6 @@ func (h *HarnessTest) RestoreNodeWithSeed(name string, extraArgs []string,
 	return node
 }
 
-// GetBestBlock makes a RPC request to miner and asserts.
-func (h *HarnessTest) GetBestBlock() (*chainhash.Hash, int32) {
-	blockHash, height, err := h.net.Miner.Client.GetBestBlock()
-	require.NoError(h, err, "failed to GetBestBlock")
-	return blockHash, height
-}
-
-// GetRecoveryInfo uses the specifies node to make a RPC call to
-// GetRecoveryInfo and asserts.
-func (h *HarnessTest) GetRecoveryInfo(
-	hn *HarnessNode) *lnrpc.GetRecoveryInfoResponse {
-
-	ctxt, cancel := context.WithTimeout(h.runCtx, DefaultTimeout)
-	defer cancel()
-
-	req := &lnrpc.GetRecoveryInfoRequest{}
-	resp, err := hn.rpc.LN.GetRecoveryInfo(ctxt, req)
-	require.NoError(h, err, "failed to GetRecoveryInfo")
-
-	return resp
-}
-
 // Shutdown shuts down the given node and asserts that no errors occur.
 func (h *HarnessTest) Shutdown(node *HarnessNode) {
 	// The process may not be in a state to always shutdown immediately, so
@@ -346,42 +312,39 @@ func (h *HarnessTest) DisconnectNodes(a, b *HarnessNode) {
 	require.NoError(h, err, "failed to disconnect nodes")
 }
 
-// DisconnectPeer calls the DisconnectPeer RPC on a given node with a specified
-// public key string and asserts there's no error.
-func (h *HarnessTest) DisconnectPeer(hn *HarnessNode,
-	pubkey string) *lnrpc.DisconnectPeerResponse {
-
-	ctxt, cancel := context.WithTimeout(h.runCtx, DefaultTimeout)
-	defer cancel()
-
-	req := &lnrpc.DisconnectPeerRequest{PubKey: pubkey}
-	resp, err := hn.rpc.LN.DisconnectPeer(ctxt, req)
-	require.NoErrorf(h, err, "failed to disconnect peer")
-
-	return resp
+// SuspendNode suspends a running node and assert.
+func (h *HarnessTest) SuspendNode(hn *HarnessNode) func() error {
+	restartFunc, err := h.net.SuspendNode(hn)
+	require.NoError(h, err)
+	return restartFunc
 }
 
-// GetInfo calls the GetInfo RPC on a given node and asserts there's no error.
-func (h *HarnessTest) GetInfo(n *HarnessNode) *lnrpc.GetInfoResponse {
-	ctxt, cancel := context.WithTimeout(h.runCtx, DefaultTimeout)
-	defer cancel()
+// RestartNode restarts a given node and asserts.
+func (h *HarnessTest) RestartNode(hn *HarnessNode,
+	chanBackups ...*lnrpc.ChanBackupSnapshot) {
 
-	info, err := n.rpc.LN.GetInfo(ctxt, &lnrpc.GetInfoRequest{})
-	require.NoErrorf(h, err, "failed to GetInfo for node: %s", n.Cfg.Name)
-
-	return info
+	err := h.net.RestartNode(hn, nil, chanBackups...)
+	require.NoErrorf(h, err, "failed to restart node %s", hn.Name())
 }
 
-// GetInfoAssertErr calls the GetInfo RPC on a given node and asserts there's
-// an error.
-func (h *HarnessTest) GetInfoAssertErr(n *HarnessNode) error {
-	ctxt, cancel := context.WithTimeout(h.runCtx, DefaultTimeout)
-	defer cancel()
+// RestartNodeNoUnlock restarts a given node without unlocking it and asserts.
+func (h *HarnessTest) RestartNodeNoUnlock(hn *HarnessNode, wait bool) {
+	err := h.net.RestartNodeNoUnlock(hn, nil, wait)
+	require.NoErrorf(h, err, "failed to restart node %s", hn.Name())
+}
 
-	_, err := n.rpc.LN.GetInfo(ctxt, &lnrpc.GetInfoRequest{})
-	require.Error(h, err, "expect an error from GetInfo")
+// RestartNodeAndRestoreDb restarts a given node with a callback to restore the
+// db.
+func (h *HarnessTest) RestartNodeAndRestoreDb(hn *HarnessNode) {
+	cb := func() error { return h.net.RestoreDb(hn) }
+	err := h.net.RestartNode(hn, cb)
+	require.NoErrorf(h, err, "failed to restart node %s", hn.Name())
+}
 
-	return err
+// StopNode stops a given node and asserts.
+func (h *HarnessTest) StopNode(hn *HarnessNode) {
+	err := h.net.StopNode(hn)
+	require.NoErrorf(h, err, "failed to stop node %s", hn.Name())
 }
 
 // SetFeeEstimate sets a fee rate to be returned from fee estimator.
@@ -525,102 +488,6 @@ func (h *HarnessTest) SendCoinsP2TR(amt btcutil.Amount, target *HarnessNode) {
 		target.Cfg.Name)
 }
 
-// SendOutputsWithoutChange uses the miner to send the given outputs using the
-// specified fee rate and returns the txid.
-func (h *HarnessTest) SendOutputsWithoutChange(
-	outputs []*wire.TxOut, feeRate btcutil.Amount) *chainhash.Hash {
-
-	txid, err := h.net.Miner.SendOutputsWithoutChange(
-		outputs, feeRate,
-	)
-	require.NoError(h, err, "failed to send output")
-
-	return txid
-}
-
-// CreateTransaction uses the miner to create a transaction using the given
-// outputs using the specified fee rate and returns the transaction.
-func (h *HarnessTest) CreateTransaction(
-	outputs []*wire.TxOut, feeRate btcutil.Amount) *wire.MsgTx {
-
-	tx, err := h.net.Miner.CreateTransaction(outputs, feeRate, false)
-	require.NoError(h, err, "failed to create transaction")
-
-	return tx
-}
-
-// OpenChannelParams houses the params to specify when opening a new channel.
-type OpenChannelParams struct {
-	// Amt is the local amount being put into the channel.
-	Amt btcutil.Amount
-
-	// PushAmt is the amount that should be pushed to the remote when the
-	// channel is opened.
-	PushAmt btcutil.Amount
-
-	// Private is a boolan indicating whether the opened channel should be
-	// private.
-	Private bool
-
-	// SpendUnconfirmed is a boolean indicating whether we can utilize
-	// unconfirmed outputs to fund the channel.
-	SpendUnconfirmed bool
-
-	// MinHtlc is the htlc_minimum_msat value set when opening the channel.
-	MinHtlc lnwire.MilliSatoshi
-
-	// RemoteMaxHtlcs is the remote_max_htlcs value set when opening the
-	// channel, restricting the number of concurrent HTLCs the remote party
-	// can add to a commitment.
-	RemoteMaxHtlcs uint16
-
-	// FundingShim is an optional funding shim that the caller can specify
-	// in order to modify the channel funding workflow.
-	FundingShim *lnrpc.FundingShim
-
-	// SatPerVByte is the amount of satoshis to spend in chain fees per
-	// virtual byte of the transaction.
-	SatPerVByte btcutil.Amount
-
-	// CommitmentType is the commitment type that should be used for the
-	// channel to be opened.
-	CommitmentType lnrpc.CommitmentType
-}
-
-type OpenChanClient lnrpc.Lightning_OpenChannelClient
-
-// OpenChannelStreamAndAssert blocks until an OpenChannel request for a channel
-// funding by alice succeeds. If it does, a stream client is returned to
-// receive events about the opening channel.
-func (h *HarnessTest) OpenChannelStreamAndAssert(a, b *HarnessNode,
-	p OpenChannelParams) OpenChanClient {
-
-	var client OpenChanClient
-	// Wait until we are able to fund a channel successfully. This wait
-	// prevents us from erroring out when trying to create a channel while
-	// the node is starting up.
-	err := wait.NoError(func() error {
-		chanOpenUpdate, err := h.net.OpenChannel(a, b, p)
-		client = chanOpenUpdate
-		return err
-	}, ChannelOpenTimeout)
-	require.NoError(h, err, "timeout when open channel between %s "+
-		"and %s", a.Name(), b.Name())
-
-	return client
-}
-
-// WaitForChannelOpen takes a open channel client and consumes the stream to
-// assert the channel is open.
-func (h *HarnessTest) WaitForChannelOpen(
-	client OpenChanClient) *lnrpc.ChannelPoint {
-
-	fundingChanPoint, err := h.net.WaitForChannelOpen(client)
-	require.NoError(h, err, "error while waiting for channel open")
-
-	return fundingChanPoint
-}
-
 // OpenPendingChannel opens a channel between the two nodes and asserts it's
 // pending.
 func (h *HarnessTest) OpenPendingChannel(from, to *HarnessNode,
@@ -732,106 +599,6 @@ func (h *HarnessTest) OpenChannelPsbt(srcNode, destNode *HarnessNode,
 	return respStream, upd.PsbtFund.Psbt
 }
 
-// ReceiveChanUpdate waits until a message is received on the stream or the
-// timeout is reached.
-func (h *HarnessTest) ReceiveChanUpdate(
-	stream OpenChanClient) *lnrpc.OpenStatusUpdate {
-
-	chanMsg := make(chan *lnrpc.OpenStatusUpdate)
-	errChan := make(chan error)
-	go func() {
-		// Consume one message. This will block until the message is
-		// received.
-		resp, err := stream.Recv()
-		if err != nil {
-			errChan <- err
-			return
-		}
-		chanMsg <- resp
-	}()
-
-	select {
-	case <-time.After(DefaultTimeout):
-		require.Fail(h, "timeout", "timeout before chan pending "+
-			"update sent")
-
-	case err := <-errChan:
-		require.Failf(h, "err from stream",
-			"received err from stream: %v", err)
-
-	case updateMsg := <-chanMsg:
-		return updateMsg
-	}
-
-	return nil
-}
-
-// AssertChannelOpen asserts that a given channel outpoint is seen by the
-// passed node.
-func (h *HarnessTest) AssertChannelOpen(n *HarnessNode,
-	chanPoint *lnrpc.ChannelPoint) {
-
-	err := n.WaitForNetworkChannelOpen(chanPoint)
-	require.NoErrorf(h, err, "%s didn't report channel", n.Cfg.Name)
-}
-
-// AssertChannelExists asserts that an active channel identified by the
-// specified channel point exists from the point-of-view of the node.
-func (h *HarnessTest) AssertChannelExists(hn *HarnessNode,
-	cp *lnrpc.ChannelPoint) *lnrpc.Channel {
-
-	var (
-		channel *lnrpc.Channel
-		err     error
-	)
-
-	err = wait.NoError(func() error {
-		channel, err = h.findChannel(hn, cp)
-		if err != nil {
-			return err
-		}
-
-		// Check whether our channel is active, exit early if it is.
-		if channel.Active {
-			return nil
-		}
-
-		return fmt.Errorf("channel point not active")
-	}, DefaultTimeout)
-
-	require.NoError(h, err, "timeout checking for channel point: %v", cp)
-
-	return channel
-}
-
-// AssertInvoiceState asserts that a given invoice has became the desired
-// state before timeout and returns the invoice found.
-func (h *HarnessTest) AssertInvoiceState(hn *HarnessNode, payHash lntypes.Hash,
-	state lnrpc.Invoice_InvoiceState) *lnrpc.Invoice {
-
-	invoice, err := hn.waitForInvoiceState(payHash, state)
-	require.NoError(h, err, "%s failed to wait invoice:%v to be state: %v",
-		hn.Name(), payHash, state)
-
-	return invoice
-}
-
-// ListChannels list the channels for the given node and asserts it's
-// successful.
-func (h *HarnessTest) ListChannels(
-	hn *HarnessNode) *lnrpc.ListChannelsResponse {
-
-	req := &lnrpc.ListChannelsRequest{}
-
-	ctxt, cancel := context.WithTimeout(h.runCtx, DefaultTimeout)
-	defer cancel()
-
-	resp, err := hn.rpc.LN.ListChannels(ctxt, req)
-	require.NoErrorf(h, err, "unable ListChannels for %s", hn.Name())
-
-	return resp
-}
-
 // MineBlocksAndAssertTx mine 'num' of blocks and check that blocks are present
 // in node blockchain. numTxs should be set to the number of transactions
 // (excluding the coinbase) we expect to be included in the first mined block.
@@ -854,15 +621,6 @@ func (h *HarnessTest) MineBlocksAndAssertTx(num uint32,
 	}
 
 	return blocks
-}
-
-// GenerateBlocks mine 'num' of blocks and returns them.
-func (h *HarnessTest) GenerateBlocks(num uint32) []*chainhash.Hash {
-	blockHashes, err := h.net.Miner.Client.Generate(num)
-	require.NoError(h, err, "unable to generate blocks")
-	require.Len(h, blockHashes, int(num), "wrong num of blocks generated")
-
-	return blockHashes
 }
 
 // MineBlocks mine 'num' of blocks and check that blocks are present in
@@ -918,86 +676,6 @@ func (h *HarnessTest) DisconnectMiner() {
 	require.NoError(h, err, "failed to disconnect miner")
 }
 
-// AssertTxInBlock asserts that a given txid can be found in the passed block.
-func (h *HarnessTest) AssertTxInBlock(block *wire.MsgBlock,
-	txid *chainhash.Hash) {
-
-	blockTxes := make([]chainhash.Hash, 0)
-
-	for _, tx := range block.Transactions {
-		sha := tx.TxHash()
-		blockTxes = append(blockTxes, sha)
-
-		if bytes.Equal(txid[:], sha[:]) {
-			return
-		}
-	}
-
-	require.Failf(h, "tx was not included in block", "tx:%v, block has:%v",
-		txid, blockTxes)
-}
-
-// AssertTxInMempool asserts a given transaction can be found in the mempool.
-func (h *HarnessTest) AssertTxInMempool(op wire.OutPoint) *wire.MsgTx {
-	var msgTx *wire.MsgTx
-
-	err := wait.NoError(func() error {
-		// We require the RPC call to be succeeded and won't wait for
-		// it as it's an unexpected behavior.
-		mempool := h.GetRawMempool()
-
-		if len(mempool) == 0 {
-			return fmt.Errorf("empty mempool")
-		}
-
-		for _, txid := range mempool {
-			// We require the RPC call to be succeeded and won't
-			// wait for it as it's an unexpected behavior.
-			tx := h.GetRawTransaction(txid)
-
-			msgTx = tx.MsgTx()
-			for _, txIn := range msgTx.TxIn {
-				if txIn.PreviousOutPoint == op {
-					return nil
-				}
-			}
-		}
-
-		return fmt.Errorf("outpoint %v not found in mempool", op)
-	}, MinerMempoolTimeout)
-
-	require.NoError(h, err, "timeout checking mempool")
-
-	return msgTx
-}
-
-// AssertNumTxsInMempool polls until finding the desired number of transactions
-// in the provided miner's mempool. It will asserrt if this number is not met
-// after the given timeout.
-func (h *HarnessTest) AssertNumTxsInMempool(n int) []*chainhash.Hash {
-	var (
-		mempool []*chainhash.Hash
-		err     error
-	)
-
-	err = wait.NoError(func() error {
-		// We require the RPC call to be succeeded and won't wait for
-		// it as it's an unexpected behavior.
-		mem := h.GetRawMempool()
-
-		if len(mem) == n {
-			mempool = mem
-			return nil
-		}
-
-		return fmt.Errorf("want %v, got %v in mempool: %v",
-			n, len(mempool), mempool)
-	}, MinerMempoolTimeout)
-	require.NoError(h, err, "assert tx in mempool timeout")
-
-	return mempool
-}
-
 // Random32Bytes generates a random 32 bytes which can be used as a pay hash,
 // preimage, etc.
 func (h *HarnessTest) Random32Bytes() []byte {
@@ -1007,57 +685,6 @@ func (h *HarnessTest) Random32Bytes() []byte {
 	require.NoErrorf(h, err, "internal error, cannot generate random bytes")
 
 	return randBuf
-}
-
-type PaymentClient routerrpc.Router_SendPaymentV2Client
-
-// SendPayment sends a payment using the given node and payment request. It
-// also asserts the payment being sent successfully.
-func (h *HarnessTest) SendPayment(hn *HarnessNode,
-	req *routerrpc.SendPaymentRequest) PaymentClient {
-
-	// SendPayment needs to have the context alive for the entire test case
-	// as the router relies on the context to propagate HTLCs. Thus we use
-	// runCtx here instead of a timeout context.
-	stream, err := hn.rpc.Router.SendPaymentV2(h.runCtx, req)
-	require.NoErrorf(h, err, "%s failed to send payment", hn.Name())
-	return stream
-}
-
-// AssertPaymentStatusFromStream takes a client stream and asserts the payment
-// is in desired status before timeout. The payment found is returned once
-// succeeded.
-func (h *HarnessTest) AssertPaymentStatusFromStream(stream PaymentClient,
-	status lnrpc.Payment_PaymentStatus) *lnrpc.Payment {
-
-	var target *lnrpc.Payment
-	err := wait.NoError(func() error {
-		// Consume one message. This will block until the
-		// message is received.
-		payment, err := stream.Recv()
-
-		// We will not wait if there's an error returned from
-		// the stream.
-		if err != nil {
-			return fmt.Errorf("payment stream "+
-				"got err: %w", err)
-		}
-
-		// Return if the desired payment state is reached.
-		if payment.Status == status {
-			target = payment
-			return nil
-		}
-
-		// Return the err so that it can be used for debugging
-		// when timeout is reached.
-		return fmt.Errorf("payment status, got %v, want %v",
-			payment.Status, status)
-	}, DefaultTimeout)
-
-	require.NoError(h, err, "timeout while waiting payment")
-
-	return target
 }
 
 // SendPaymentAndAssertStatus sends a payment from the passed node and asserts
@@ -1091,98 +718,6 @@ func (h *HarnessTest) SendPaymentAssertFail(hn *HarnessNode,
 	return payment
 }
 
-// assertNodeNumChannels polls the provided node's list channels rpc until it
-// reaches the desired number of total channels.
-func (h *HarnessTest) AssertNodeNumChannels(hn *HarnessNode, numChannels int) {
-	err := wait.NoError(func() error {
-		// We require the RPC call to be succeeded and won't wait for
-		// it as it's an unexpected behavior.
-		chanInfo := h.ListChannels(hn)
-
-		// Return true if the query returned the expected number of
-		// channels.
-		num := len(chanInfo.Channels)
-		if num != numChannels {
-			return fmt.Errorf("expected %v channels, got %v",
-				numChannels, num)
-		}
-		return nil
-	}, DefaultTimeout)
-
-	require.NoError(h, err, "timeout checking node's num of channels")
-}
-
-// AssertNumActiveHtlcs asserts that a given number of HTLCs are seen in the
-// node's channels.
-func (h *HarnessTest) AssertNumActiveHtlcs(hn *HarnessNode, num int) {
-	err := wait.NoError(func() error {
-		// We require the RPC call to be succeeded and won't wait for
-		// it as it's an unexpected behavior.
-		nodeChans := h.ListChannels(hn)
-
-		total := 0
-		for _, channel := range nodeChans.Channels {
-			total += len(channel.PendingHtlcs)
-		}
-		if total != num {
-			return fmt.Errorf("expected %v HTLCs, got %v",
-				num, total)
-		}
-
-		return nil
-	}, DefaultTimeout)
-
-	require.NoErrorf(h, err, "%s assert active htlcs timed out", hn.Name())
-}
-
-// AssertActiveHtlcs makes sure all the passed nodes have the _exact_ HTLCs
-// matching payHashes on _all_ their channels.
-func (h *HarnessTest) AssertActiveHtlcs(hn *HarnessNode, payHashes ...[]byte) {
-	checkPayHashesMatch := func() error {
-		// We require the RPC call to be succeeded and won't wait for
-		// it as it's an unexpected behavior.
-		nodeChans := h.ListChannels(hn)
-
-		for _, ch := range nodeChans.Channels {
-			// Record all payment hashes active for this channel.
-			htlcHashes := make(map[string]struct{})
-
-			for _, htlc := range ch.PendingHtlcs {
-				h := hex.EncodeToString(htlc.HashLock)
-				_, ok := htlcHashes[h]
-				if ok {
-					return fmt.Errorf("duplicate HashLock")
-				}
-				htlcHashes[h] = struct{}{}
-			}
-
-			// Channel should have exactly the payHashes active.
-			if len(payHashes) != len(htlcHashes) {
-				return fmt.Errorf("node [%s:%x] had %v "+
-					"htlcs active, expected %v",
-					hn.Name(), hn.PubKey[:],
-					len(htlcHashes), len(payHashes))
-			}
-
-			// Make sure all the payHashes are active.
-			for _, payHash := range payHashes {
-				h := hex.EncodeToString(payHash)
-				if _, ok := htlcHashes[h]; ok {
-					continue
-				}
-				return fmt.Errorf("node [%s:%x] didn't have: "+
-					"the payHash %v active", hn.Name(),
-					hn.PubKey[:], h)
-			}
-		}
-
-		return nil
-	}
-
-	err := wait.NoError(checkPayHashesMatch, DefaultTimeout)
-	require.NoError(h, err)
-}
-
 // OutPointFromChannelPoint creates an outpoint from a given channel point.
 func (h *HarnessTest) OutPointFromChannelPoint(
 	cp *lnrpc.ChannelPoint) wire.OutPoint {
@@ -1192,182 +727,6 @@ func (h *HarnessTest) OutPointFromChannelPoint(
 		Hash:  *txid,
 		Index: cp.OutputIndex,
 	}
-}
-
-// GetNumTxsFromMempool polls until finding the desired number of transactions
-// in the miner's mempool and returns the full transactions to the caller.
-func (h *HarnessTest) GetNumTxsFromMempool(n int) []*wire.MsgTx {
-	txids := h.AssertNumTxsInMempool(n)
-
-	var txes []*wire.MsgTx
-	for _, txid := range txids {
-		tx := h.GetRawTransaction(txid)
-		txes = append(txes, tx.MsgTx())
-	}
-	return txes
-}
-
-// GetRawTransaction makes a RPC call to the miner's GetRawTransaction and
-// asserts.
-func (h *HarnessTest) GetRawTransaction(txid *chainhash.Hash) *btcutil.Tx {
-	tx, err := h.net.Miner.Client.GetRawTransaction(txid)
-	require.NoErrorf(h, err, "failed to get raw tx: %v", txid)
-	return tx
-}
-
-// GetRawMempool makes a RPC call to the miner's GetRawMempool and
-// asserts.
-func (h *HarnessTest) GetRawMempool() []*chainhash.Hash {
-	mempool, err := h.net.Miner.Client.GetRawMempool()
-	require.NoError(h, err, "unable to get mempool")
-	return mempool
-}
-
-// assertAllTxesSpendFrom asserts that all txes in the list spend from the
-// given tx.
-func (h *HarnessTest) AssertAllTxesSpendFrom(txes []*wire.MsgTx,
-	prevTxid chainhash.Hash) {
-
-	for _, tx := range txes {
-		if tx.TxIn[0].PreviousOutPoint.Hash != prevTxid {
-			require.Failf(h, "", "tx %v did not spend from %v",
-				tx.TxHash(), prevTxid)
-		}
-	}
-}
-
-// AssertTxSpendFrom asserts that a given tx is spent from a previous tx.
-// tx.
-func (h *HarnessTest) AssertTxSpendFrom(tx *wire.MsgTx,
-	prevTxid chainhash.Hash) {
-
-	if tx.TxIn[0].PreviousOutPoint.Hash != prevTxid {
-		require.Failf(h, "", "tx %v did not spend from %v",
-			tx.TxHash(), prevTxid)
-	}
-}
-
-// GetPendingChannels makes a RPC request tc PendingChannels and asserts
-// there's no error.
-func (h *HarnessTest) GetPendingChannels(
-	hn *HarnessNode) *lnrpc.PendingChannelsResponse {
-
-	ctxt, cancel := context.WithTimeout(h.runCtx, DefaultTimeout)
-	defer cancel()
-
-	var (
-		resp *lnrpc.PendingChannelsResponse
-		err  error
-	)
-
-	waitErr := wait.NoError(func() error {
-		pendingChansRequest := &lnrpc.PendingChannelsRequest{}
-		resp, err = hn.rpc.LN.PendingChannels(
-			ctxt, pendingChansRequest,
-		)
-
-		// TODO(yy): the RPC call should not return an error. If an
-		// error is returned, it indicates there's something wrong
-		// regarding how we prepare the channel info. This could happen
-		// when the channel happens to be transitioning states. We log
-		// it here, but this needs to be fixed in the RPC server.
-		if err != nil {
-			h.Logf("%s PendingChannels got err: %v", hn.Name(), err)
-		}
-		return err
-	}, DefaultTimeout)
-
-	require.NoError(h, waitErr, "failed to get pending channels for %s",
-		hn.Name())
-
-	return resp
-}
-
-type PendingForceClose *lnrpc.PendingChannelsResponse_ForceClosedChannel
-
-// AssertChannelPendingForceClose asserts that the given channel found in the
-// node is pending force close. Returns the PendingForceClose if found.
-func (h *HarnessTest) AssertChannelPendingForceClose(hn *HarnessNode,
-	chanPoint *lnrpc.ChannelPoint) PendingForceClose {
-
-	var target PendingForceClose
-
-	op := h.OutPointFromChannelPoint(chanPoint)
-
-	err := wait.NoError(func() error {
-		resp := h.GetPendingChannels(hn)
-
-		forceCloseChans := resp.PendingForceClosingChannels
-		for _, ch := range forceCloseChans {
-			if ch.Channel.ChannelPoint == op.String() {
-				target = ch
-				return nil
-			}
-		}
-
-		return fmt.Errorf("%v: channel %s not found in pending "+
-			"force close", hn.Name(), chanPoint)
-	}, DefaultTimeout)
-	require.NoError(h, err, "assert pending force close timed out")
-
-	return target
-}
-
-type WaitingCloseChannel *lnrpc.PendingChannelsResponse_WaitingCloseChannel
-
-// AssertChannelWaitingClose asserts that the given channel found in the node
-// is waiting close. Returns the WaitingCloseChannel if found.
-func (h *HarnessTest) AssertChannelWaitingClose(hn *HarnessNode,
-	chanPoint *lnrpc.ChannelPoint) WaitingCloseChannel {
-
-	var target WaitingCloseChannel
-
-	op := h.OutPointFromChannelPoint(chanPoint)
-
-	err := wait.NoError(func() error {
-		resp := h.GetPendingChannels(hn)
-
-		for _, waitingClose := range resp.WaitingCloseChannels {
-			if waitingClose.Channel.ChannelPoint == op.String() {
-				target = waitingClose
-				return nil
-			}
-		}
-
-		return fmt.Errorf("%v: channel %s not found in waiting close",
-			hn.Name(), op)
-	}, DefaultTimeout)
-	require.NoError(h, err, "assert channel waiting close timed out")
-
-	return target
-}
-
-// AssertNumPendingCloseChannels checks that a PendingChannels response from
-// the node reports the expected number of pending force channels and waiting
-// close channels.
-func (h *HarnessTest) AssertNumPendingCloseChannels(hn *HarnessNode,
-	expWaitingClose, expPendingForceClose int) {
-
-	err := wait.NoError(func() error {
-		resp := h.GetPendingChannels(hn)
-
-		n := len(resp.WaitingCloseChannels)
-		if n != expWaitingClose {
-			return fmt.Errorf("expected to find %d channels "+
-				"waiting close, found %d", expWaitingClose, n)
-		}
-
-		n = len(resp.PendingForceClosingChannels)
-		if n != expPendingForceClose {
-			return fmt.Errorf("expected to find %d channel "+
-				"pending force close, found %d",
-				expPendingForceClose, n)
-		}
-
-		return nil
-	}, DefaultTimeout)
-
-	require.NoErrorf(h, err, "assert pending channels timeout")
 }
 
 // closeChannel attempts to close a channel identified by the passed channel
@@ -1445,71 +804,6 @@ func (h *HarnessTest) CloseReorgedChannel(hn *HarnessNode,
 	return h.assertChannelClosed(
 		hn, fundingChanPoint, false, closeUpdates,
 	)
-}
-
-type CloseChanClient lnrpc.Lightning_CloseChannelClient
-
-// CloseChannelStreamAndAssert blocks until an CloseChannel request for a
-// given channel point succeeds. If it does, a stream client is returned
-// to receive events about the closing channel.
-func (h *HarnessTest) CloseChannelStreamAndAssert(hn *HarnessNode,
-	chanPoint *lnrpc.ChannelPoint, force bool) CloseChanClient {
-
-	var (
-		client CloseChanClient
-		err    error
-	)
-
-	req := &lnrpc.CloseChannelRequest{
-		ChannelPoint: chanPoint,
-		Force:        force,
-	}
-
-	err = wait.NoError(func() error {
-		// Use runCtx here instead of a timeout context to keep the
-		// client alive for the entire test case.
-		client, err = hn.rpc.LN.CloseChannel(h.runCtx, req)
-		if err != nil {
-			return fmt.Errorf("unable to close channel for node %s"+
-				" with channel point %s", hn.Name(), chanPoint)
-		}
-		return nil
-	}, ChannelCloseTimeout)
-
-	require.NoError(h, err, "timeout closing channel")
-
-	return client
-}
-
-// AssertChannelPolicyUpdate checks that the required policy update has
-// happened on the given node.
-func (h *HarnessTest) AssertChannelPolicyUpdate(node *HarnessNode,
-	advertisingNode string, policy *lnrpc.RoutingPolicy,
-	chanPoint *lnrpc.ChannelPoint, includeUnannounced bool) {
-
-	require.NoError(
-		h, node.WaitForChannelPolicyUpdate(
-			advertisingNode, policy,
-			chanPoint, includeUnannounced,
-		), "error while waiting for channel update",
-	)
-}
-
-// WaitForChannelClose will consume a close channel stream and asserts the
-// channel is closed, return the closing txid if succeeded.
-func (h *HarnessTest) WaitForChannelClose(
-	closeUpdates CloseChanClient) *chainhash.Hash {
-
-	var txid *chainhash.Hash
-
-	err := wait.NoError(func() error {
-		closingTxid, err := h.net.WaitForChannelClose(closeUpdates)
-		txid = closingTxid
-		return err
-	}, ChannelCloseTimeout)
-	require.NoError(h, err, "error while waiting for channel close")
-
-	return txid
 }
 
 // assertChannelClosed asserts that the channel is properly cleaned up after
@@ -1629,167 +923,6 @@ func (h *HarnessTest) getChannelPolicies(hn *HarnessNode,
 	return policies
 }
 
-// ListPeers makes a RPC call to the node's ListPeers and asserts.
-func (h *HarnessTest) ListPeers(hn *HarnessNode) *lnrpc.ListPeersResponse {
-	ctxt, cancel := context.WithTimeout(h.runCtx, DefaultTimeout)
-	defer cancel()
-
-	resp, err := hn.rpc.LN.ListPeers(ctxt, &lnrpc.ListPeersRequest{})
-	require.NoErrorf(h, err, "%s ListPeers failed with: %v",
-		hn.Name(), err)
-
-	return resp
-}
-
-// assertPeerConnected asserts that the given node b is connected to a.
-func (h *HarnessTest) assertPeerConnected(a, b *HarnessNode) {
-	err := wait.NoError(func() error {
-		// We require the RPC call to be succeeded and won't wait for
-		// it as it's an unexpected behavior.
-		resp := h.ListPeers(a)
-
-		// If node B is seen in the ListPeers response from node A,
-		// then we can return true as the connection has been fully
-		// established.
-		for _, peer := range resp.Peers {
-			if peer.PubKey == b.PubKeyStr {
-				return nil
-			}
-		}
-
-		return fmt.Errorf("%s not found in %s's ListPeers",
-			b.Name(), a.Name())
-	}, DefaultTimeout)
-
-	require.NoError(h, err, "unable to connect %s to %s, got error: "+
-		"peers not connected within %v seconds",
-		a.Name(), b.Name(), DefaultTimeout)
-}
-
-// AssertConnected asserts that two peers are connected.
-func (h *HarnessTest) AssertConnected(a, b *HarnessNode) {
-	h.assertPeerConnected(a, b)
-	h.assertPeerConnected(b, a)
-}
-
-// assertPeerNotConnected asserts that the given node b is not connected to a.
-func (h *HarnessTest) assertPeerNotConnected(a, b *HarnessNode) {
-	err := wait.NoError(func() error {
-		// We require the RPC call to be succeeded and won't wait for
-		// it as it's an unexpected behavior.
-		resp := h.ListPeers(a)
-
-		// If node B is seen in the ListPeers response from node A,
-		// then we return false as the connection has been fully
-		// established.
-		for _, peer := range resp.Peers {
-			if peer.PubKey == b.PubKeyStr {
-				return fmt.Errorf("peers %s and %s still "+
-					"connected", a.Name(), b.Name())
-			}
-		}
-		return nil
-	}, DefaultTimeout)
-
-	require.NoError(h, err, "timeout checking peers not connected")
-}
-
-// AssertNotConnected asserts that two peers are not connected.
-func (h *HarnessTest) AssertNotConnected(a, b *HarnessNode) {
-	h.assertPeerNotConnected(a, b)
-	h.assertPeerNotConnected(b, a)
-}
-
-// AddHoldInvoice adds a hold invoice for the given node and asserts.
-func (h *HarnessTest) AddHoldInvoice(
-	req *invoicesrpc.AddHoldInvoiceRequest,
-	hn *HarnessNode) *invoicesrpc.AddHoldInvoiceResp {
-
-	ctxt, cancel := context.WithTimeout(h.runCtx, DefaultTimeout)
-	defer cancel()
-
-	invoice, err := hn.rpc.Invoice.AddHoldInvoice(ctxt, req)
-	require.NoError(h, err)
-
-	return invoice
-}
-
-// AddInvoice adds a invoice for the given node and asserts.
-func (h *HarnessTest) AddInvoice(req *lnrpc.Invoice,
-	hn *HarnessNode) *lnrpc.AddInvoiceResponse {
-
-	ctxt, cancel := context.WithTimeout(h.runCtx, DefaultTimeout)
-	defer cancel()
-
-	invoice, err := hn.rpc.LN.AddInvoice(ctxt, req)
-	require.NoError(h, err)
-
-	return invoice
-}
-
-// SuspendNode suspends a running node and assert.
-func (h *HarnessTest) SuspendNode(hn *HarnessNode) func() error {
-	restartFunc, err := h.net.SuspendNode(hn)
-	require.NoError(h, err)
-	return restartFunc
-}
-
-// SettleInvoice settles a given invoice and asserts.
-func (h *HarnessTest) SettleInvoice(hn *HarnessNode,
-	preimage []byte) *invoicesrpc.SettleInvoiceResp {
-
-	ctxt, cancel := context.WithTimeout(h.runCtx, DefaultTimeout)
-	defer cancel()
-
-	req := &invoicesrpc.SettleInvoiceMsg{Preimage: preimage}
-	resp, err := hn.rpc.Invoice.SettleInvoice(ctxt, req)
-	require.NoError(h, err)
-
-	return resp
-}
-
-// CancelInvoice cancels a given invoice and asserts.
-func (h *HarnessTest) CancelInvoice(hn *HarnessNode,
-	preimage []byte) *invoicesrpc.CancelInvoiceResp {
-
-	ctxt, cancel := context.WithTimeout(h.runCtx, DefaultTimeout)
-	defer cancel()
-
-	req := &invoicesrpc.CancelInvoiceMsg{PaymentHash: preimage}
-	resp, err := hn.rpc.Invoice.CancelInvoice(ctxt, req)
-	require.NoError(h, err)
-
-	return resp
-}
-
-// ListInvoices list the node's invoice using the request and asserts.
-func (h *HarnessTest) ListInvoices(hn *HarnessNode) *lnrpc.ListInvoiceResponse {
-	ctxt, cancel := context.WithTimeout(h.runCtx, DefaultTimeout)
-	defer cancel()
-
-	req := &lnrpc.ListInvoiceRequest{}
-	resp, err := hn.rpc.LN.ListInvoices(ctxt, req)
-	require.NoErrorf(h, err, "list invoice failed")
-
-	return resp
-}
-
-// ListPayments lists the node's payments and asserts.
-func (h *HarnessTest) ListPayments(hn *HarnessNode,
-	includeIncomplete bool) *lnrpc.ListPaymentsResponse {
-
-	ctxt, cancel := context.WithTimeout(h.runCtx, DefaultTimeout)
-	defer cancel()
-
-	req := &lnrpc.ListPaymentsRequest{
-		IncludeIncomplete: includeIncomplete,
-	}
-	resp, err := hn.rpc.LN.ListPayments(ctxt, req)
-	require.NoError(h, err, "failed to list payments")
-
-	return resp
-}
-
 // findPayment queries the payment from the node's ListPayments which matches
 // the specified preimage hash.
 func (h *HarnessTest) findPayment(hn *HarnessNode,
@@ -1805,96 +938,8 @@ func (h *HarnessTest) findPayment(hn *HarnessNode,
 		return p
 	}
 
-	require.Fail(h, "payment not found")
+	require.Fail(h, "payment: %v not found", payHash)
 	return nil
-}
-
-// AssertPaymentStatus asserts that the given node list a payment with the
-// given preimage has the expected status. It also checks that the payment has
-// the expected preimage, which is empty when it's not settled and matches the
-// given preimage when it's succeeded.
-func (h *HarnessTest) AssertPaymentStatus(hn *HarnessNode,
-	preimage lntypes.Preimage,
-	status lnrpc.Payment_PaymentStatus) *lnrpc.Payment {
-
-	var target *lnrpc.Payment
-
-	err := wait.NoError(func() error {
-		p := h.findPayment(hn, preimage)
-		if status == p.Status {
-			target = p
-			return nil
-		}
-		return fmt.Errorf("payment status not match, want %s "+"got %s",
-			status, p.Status)
-	}, DefaultTimeout)
-	require.NoError(h, err, "timeout checking payment status")
-
-	switch status {
-	// If this expected status is SUCCEEDED, we expect the final
-	// preimage.
-	case lnrpc.Payment_SUCCEEDED:
-		require.Equal(h, preimage.String(), target.PaymentPreimage,
-			"preimage not match")
-
-	// Otherwise we expect an all-zero preimage.
-	default:
-		require.Equal(h, (lntypes.Preimage{}).String(),
-			target.PaymentPreimage, "expected zero preimage")
-	}
-
-	return target
-}
-
-// AssertHTLCStage asserts the HTLC is in the expected stage or timeout.
-func (h *HarnessTest) AssertHTLCStage(hn *HarnessNode, stage uint32) {
-	checkStage := func() error {
-		pendingChanResp := h.GetPendingChannels(hn)
-		if len(pendingChanResp.PendingForceClosingChannels) == 0 {
-			return fmt.Errorf("zero pending force closing channels")
-		}
-
-		ch := pendingChanResp.PendingForceClosingChannels[0]
-		if ch.LimboBalance == 0 {
-			return fmt.Errorf("zero balance")
-		}
-
-		if len(ch.PendingHtlcs) != 1 {
-			return fmt.Errorf("got %d pending htlcs, want 1",
-				len(ch.PendingHtlcs))
-		}
-		if stage != ch.PendingHtlcs[0].Stage {
-			return fmt.Errorf("got stage: %v, want stage: %v",
-				ch.PendingHtlcs[0].Stage, stage)
-		}
-
-		return nil
-	}
-
-	require.NoErrorf(h, wait.NoError(checkStage, DefaultTimeout),
-		"timeout waiting for htlc stage")
-}
-
-// NewMinerAddress creates a new address for the miner and asserts.
-func (h *HarnessTest) NewMinerAddress() btcutil.Address {
-	addr, err := h.net.Miner.NewAddress()
-	require.NoError(h, err, "failed to create new miner address")
-	return addr
-}
-
-// GetTransactions makes a RPC call to GetTransactions and asserts.
-func (h *HarnessTest) GetTransactions(
-	hn *HarnessNode) *lnrpc.TransactionDetails {
-
-	ctxt, cancel := context.WithTimeout(h.runCtx, DefaultTimeout)
-	defer cancel()
-
-	resp, err := hn.rpc.LN.GetTransactions(
-		ctxt, &lnrpc.GetTransactionsRequest{},
-	)
-	require.NoError(h, err, "failed to GetTransactions for %s", hn.Name())
-
-	return resp
 }
 
 // txStr returns the string representation of the channel's funding transaction.
@@ -1908,399 +953,6 @@ func txStr(chanPoint *lnrpc.ChannelPoint) string {
 		Index: chanPoint.OutputIndex,
 	}
 	return cp.String()
-}
-
-// GetWalletBalance makes a RPC call to WalletBalance and asserts.
-func (h *HarnessTest) GetWalletBalance(
-	hn *HarnessNode) *lnrpc.WalletBalanceResponse {
-
-	ctxt, cancel := context.WithTimeout(h.runCtx, DefaultTimeout)
-	defer cancel()
-
-	req := &lnrpc.WalletBalanceRequest{}
-	resp, err := hn.rpc.LN.WalletBalance(ctxt, req)
-	require.NoError(h, err, "failed to get wallet balance for node %s",
-		hn.Name())
-
-	return resp
-}
-
-// ListUnspent makes a RPC call to ListUnspent and asserts.
-func (h *HarnessTest) ListUnspent(hn *HarnessNode, account string,
-	max, min int32) *lnrpc.ListUnspentResponse {
-
-	ctxt, cancel := context.WithTimeout(h.runCtx, DefaultTimeout)
-	defer cancel()
-
-	req := &lnrpc.ListUnspentRequest{
-		Account:  account,
-		MaxConfs: max,
-		MinConfs: min,
-	}
-	resp, err := hn.rpc.LN.ListUnspent(ctxt, req)
-	require.NoError(h, err, "failed to list utxo for node %s", hn.Name())
-
-	return resp
-}
-
-// NewAddress makes a RPC call to NewAddress and asserts.
-func (h *HarnessTest) NewAddress(hn *HarnessNode,
-	addrType lnrpc.AddressType) *lnrpc.NewAddressResponse {
-
-	ctxt, cancel := context.WithTimeout(h.runCtx, DefaultTimeout)
-	defer cancel()
-
-	req := &lnrpc.NewAddressRequest{Type: addrType}
-	resp, err := hn.rpc.LN.NewAddress(ctxt, req)
-	require.NoError(h, err, "failed to create new address for node %s",
-		hn.Name())
-
-	return resp
-}
-
-// NewAddressWithAccount makes a RPC call to NewAddress and asserts.
-func (h *HarnessTest) NewAddressWithAccount(hn *HarnessNode,
-	addrType lnrpc.AddressType, account string) *lnrpc.NewAddressResponse {
-
-	ctxt, cancel := context.WithTimeout(h.runCtx, DefaultTimeout)
-	defer cancel()
-
-	req := &lnrpc.NewAddressRequest{Type: addrType, Account: account}
-	resp, err := hn.rpc.LN.NewAddress(ctxt, req)
-	require.NoError(h, err, "failed to create new address for node %s",
-		hn.Name())
-
-	return resp
-}
-
-// SendCoinFromNode sends a given amount of money to the specified address from
-// the passed node.
-func (h *HarnessTest) SendCoinFromNode(hn *HarnessNode,
-	req *lnrpc.SendCoinsRequest) *lnrpc.SendCoinsResponse {
-
-	ctxt, cancel := context.WithTimeout(h.runCtx, DefaultTimeout)
-	defer cancel()
-
-	resp, err := hn.rpc.LN.SendCoins(ctxt, req)
-	require.NoError(h, err, "node %s failed to send coins to address %s",
-		hn.Name(), req.Addr)
-
-	return resp
-}
-
-// SendCoinFromNodeErr sends a given amount of money to the specified address
-// from the passed node and asserts an error has returned.
-func (h *HarnessTest) SendCoinFromNodeErr(hn *HarnessNode,
-	req *lnrpc.SendCoinsRequest) {
-
-	ctxt, cancel := context.WithTimeout(h.runCtx, DefaultTimeout)
-	defer cancel()
-
-	_, err := hn.rpc.LN.SendCoins(ctxt, req)
-	require.Error(h, err, "node %s didn't not return an error", hn.Name())
-}
-
-// GetChannelBalance gets the channel balance and asserts.
-func (h *HarnessTest) GetChannelBalance(
-	hn *HarnessNode) *lnrpc.ChannelBalanceResponse {
-
-	ctxt, cancel := context.WithTimeout(h.runCtx, DefaultTimeout)
-	defer cancel()
-
-	req := &lnrpc.ChannelBalanceRequest{}
-	resp, err := hn.rpc.LN.ChannelBalance(ctxt, req)
-
-	require.NoError(h, err, "unable to get balance for node %s", hn.Name())
-	return resp
-}
-
-// AssertChannelBalanceResp makes a ChannelBalance request and checks the
-// returned response matches the expected.
-func (h *HarnessTest) AssertChannelBalanceResp(hn *HarnessNode,
-	expected *lnrpc.ChannelBalanceResponse) { // nolint:interfacer
-
-	resp := h.GetChannelBalance(hn)
-	require.True(h, proto.Equal(expected, resp), "balance is incorrect "+
-		"got: %v, want: %v", resp, expected,
-	)
-}
-
-// GetChannelCommitType retrieves the active channel commitment type for the
-// given chan point.
-func (h *HarnessTest) GetChannelCommitType(hn *HarnessNode,
-	chanPoint *lnrpc.ChannelPoint) lnrpc.CommitmentType {
-
-	channels := h.ListChannels(hn)
-	for _, c := range channels.Channels {
-		if c.ChannelPoint == txStr(chanPoint) {
-			return c.CommitmentType
-		}
-	}
-
-	require.Fail(h, "get channel commit type failed", "channel point %v "+
-		"not found", chanPoint)
-
-	return 0
-}
-
-// DeriveKey makes a RPC call to the DeriveKey and asserts.
-func (h *HarnessTest) DeriveKey(hn *HarnessNode,
-	kl *signrpc.KeyLocator) *signrpc.KeyDescriptor {
-
-	ctxt, cancel := context.WithTimeout(h.runCtx, DefaultTimeout)
-	defer cancel()
-
-	key, err := hn.rpc.WalletKit.DeriveKey(ctxt, kl)
-	require.NoError(h, err, "failed to derive key for node %s", hn.Name())
-
-	return key
-}
-
-// DeriveNextKey makes a RPC call to the DeriveNextKey and asserts.
-func (h *HarnessTest) DeriveNextKey(hn *HarnessNode,
-	req *walletrpc.KeyReq) *signrpc.KeyDescriptor {
-
-	ctxt, cancel := context.WithTimeout(h.runCtx, DefaultTimeout)
-	defer cancel()
-
-	key, err := hn.rpc.WalletKit.DeriveNextKey(ctxt, req)
-	require.NoError(h, err, "failed to derive next key for node %s",
-		hn.Name())
-
-	return key
-}
-
-// FundingStateStep makes a RPC call to FundingStateStep and asserts.
-func (h *HarnessTest) FundingStateStep(hn *HarnessNode,
-	msg *lnrpc.FundingTransitionMsg) *lnrpc.FundingStateStepResp {
-
-	ctxt, cancel := context.WithTimeout(h.runCtx, DefaultTimeout)
-	defer cancel()
-
-	resp, err := hn.rpc.LN.FundingStateStep(ctxt, msg)
-	require.NoError(h, err, "failed to step state")
-
-	return resp
-}
-
-// FundingStateStepAssertErr makes a RPC call to FundingStateStep and asserts
-// there's an error.
-func (h *HarnessTest) FundingStateStepAssertErr(hn *HarnessNode,
-	msg *lnrpc.FundingTransitionMsg) {
-
-	ctxt, cancel := context.WithTimeout(h.runCtx, DefaultTimeout)
-	defer cancel()
-
-	_, err := hn.rpc.LN.FundingStateStep(ctxt, msg)
-	require.Error(h, err, "duplicate pending channel ID funding shim "+
-		"registration should trigger an error")
-}
-
-// AssertNumOpenChannelsPending asserts that a pair of nodes have the expected
-// number of pending channels between them.
-func (h *HarnessTest) AssertNumOpenChannelsPending(alice, bob *HarnessNode,
-	expected int) {
-
-	err := wait.NoError(func() error {
-		aliceChans := h.GetPendingChannels(alice)
-		bobChans := h.GetPendingChannels(bob)
-
-		aliceNumChans := len(aliceChans.PendingOpenChannels)
-		bobNumChans := len(bobChans.PendingOpenChannels)
-
-		aliceStateCorrect := aliceNumChans == expected
-		if !aliceStateCorrect {
-			return fmt.Errorf("number of pending channels for "+
-				"alice incorrect. expected %v, got %v",
-				expected, aliceNumChans)
-		}
-
-		bobStateCorrect := bobNumChans == expected
-		if !bobStateCorrect {
-			return fmt.Errorf("number of pending channels for bob "+
-				"incorrect. expected %v, got %v", expected,
-				bobNumChans)
-		}
-
-		return nil
-	}, DefaultTimeout)
-
-	require.NoError(h, err, "num of pending open channels not match")
-}
-
-// AbandonChannel makes a RPC call to AbandonChannel and asserts.
-func (h *HarnessTest) AbandonChannel(hn *HarnessNode,
-	req *lnrpc.AbandonChannelRequest) *lnrpc.AbandonChannelResponse {
-
-	ctxt, cancel := context.WithTimeout(h.runCtx, DefaultTimeout)
-	defer cancel()
-
-	resp, err := hn.rpc.LN.AbandonChannel(ctxt, req)
-	require.NoError(h, err, "abandon channel failed")
-
-	return resp
-}
-
-// RestartNode restarts a given node and asserts.
-func (h *HarnessTest) RestartNode(hn *HarnessNode,
-	chanBackups ...*lnrpc.ChanBackupSnapshot) {
-
-	err := h.net.RestartNode(hn, nil, chanBackups...)
-	require.NoErrorf(h, err, "failed to restart node %s", hn.Name())
-}
-
-// RestartNodeNoUnlock restarts a given node without unlocking it and asserts.
-func (h *HarnessTest) RestartNodeNoUnlock(hn *HarnessNode, wait bool) {
-	err := h.net.RestartNodeNoUnlock(hn, nil, wait)
-	require.NoErrorf(h, err, "failed to restart node %s", hn.Name())
-}
-
-// RestartNodeAndRestoreDb restarts a given node with a callback to restore the
-// db.
-func (h *HarnessTest) RestartNodeAndRestoreDb(hn *HarnessNode) {
-	cb := func() error { return h.net.RestoreDb(hn) }
-	err := h.net.RestartNode(hn, cb)
-	require.NoErrorf(h, err, "failed to restart node %s", hn.Name())
-}
-
-// StopNode stops a given node and asserts.
-func (h *HarnessTest) StopNode(hn *HarnessNode) {
-	err := h.net.StopNode(hn)
-	require.NoErrorf(h, err, "failed to stop node %s", hn.Name())
-}
-
-// AssertTxAtHeight gets all of the transactions that a node's wallet has a
-// record of at the target height, and finds and returns the tx with the target
-// txid, failing if it is not found.
-func (h *HarnessTest) AssertTxAtHeight(hn *HarnessNode, height int32,
-	txid string) *lnrpc.Transaction {
-
-	ctxt, cancel := context.WithTimeout(h.runCtx, DefaultTimeout)
-	defer cancel()
-
-	txns, err := hn.rpc.LN.GetTransactions(
-		ctxt, &lnrpc.GetTransactionsRequest{
-			StartHeight: height,
-			EndHeight:   height,
-		},
-	)
-	require.NoError(h, err, "could not get transactions")
-
-	for _, tx := range txns.Transactions {
-		if tx.TxHash == txid {
-			return tx
-		}
-	}
-
-	require.Failf(h, "fail to find tx", "tx:%v not found at height:%v",
-		txid, height)
-
-	return nil
-}
-
-// BatchOpenChannel makes a RPC call to BatchOpenChannel and asserts.
-func (h *HarnessTest) BatchOpenChannel(hn *HarnessNode,
-	req *lnrpc.BatchOpenChannelRequest) *lnrpc.BatchOpenChannelResponse {
-
-	ctxt, cancel := context.WithTimeout(h.runCtx, DefaultTimeout)
-	defer cancel()
-
-	resp, err := hn.rpc.LN.BatchOpenChannel(ctxt, req)
-	require.NoError(h, err, "failed to batch open channel")
-
-	return resp
-}
-
-// BatchOpenChannelAssertErr makes a RPC call to BatchOpenChannel and asserts
-// there's an error returned.
-func (h *HarnessTest) BatchOpenChannelAssertErr(hn *HarnessNode,
-	req *lnrpc.BatchOpenChannelRequest) error {
-
-	ctxt, cancel := context.WithTimeout(h.runCtx, DefaultTimeout)
-	defer cancel()
-
-	_, err := hn.rpc.LN.BatchOpenChannel(ctxt, req)
-	require.Error(h, err, "expecte batch open channel fail")
-
-	return err
-}
-
-// AssertChannelPolicy asserts that the passed node's known channel policy for
-// the passed chanPoint is consistent with the expected policy values.
-func (h *HarnessTest) AssertChannelPolicy(hn *HarnessNode,
-	advertisingNode string, expectedPolicy *lnrpc.RoutingPolicy,
-	chanPoints ...*lnrpc.ChannelPoint) {
-
-	policies := h.getChannelPolicies(hn, advertisingNode, chanPoints...)
-	for _, policy := range policies {
-		err := CheckChannelPolicy(policy, expectedPolicy)
-		require.NoError(h, err, "check policy failed")
-	}
-}
-
-// QueryRoutes makes a RPC call to QueryRoutes and asserts.
-func (h *HarnessTest) QueryRoutes(hn *HarnessNode,
-	req *lnrpc.QueryRoutesRequest) *lnrpc.QueryRoutesResponse {
-
-	ctxt, cancel := context.WithTimeout(h.runCtx, DefaultTimeout)
-	defer cancel()
-
-	routes, err := hn.rpc.LN.QueryRoutes(ctxt, req)
-	require.NoError(h, err, "failed to query routes")
-
-	return routes
-}
-
-// SendToRoute makes a RPC call to SendToRoute and asserts.
-func (h *HarnessTest) SendToRoute(
-	hn *HarnessNode) lnrpc.Lightning_SendToRouteClient {
-
-	// SendToRoute needs to have the context alive for the entire test case
-	// as the returned client will be used for send and receive payment
-	// stream. Thus we use runCtx here instead of a timeout context.
-	client, err := hn.rpc.LN.SendToRoute(h.runCtx) // nolint:staticcheck
-	require.NoError(h, err, "failed to send to route")
-
-	return client
-}
-
-// SendToRouteSync makes a RPC call to SendToRouteSync and asserts.
-func (h *HarnessTest) SendToRouteSync(hn *HarnessNode,
-	req *lnrpc.SendToRouteRequest) *lnrpc.SendResponse {
-
-	ctxt, cancel := context.WithTimeout(h.runCtx, DefaultTimeout)
-	defer cancel()
-
-	resp, err := hn.rpc.LN.SendToRouteSync(ctxt, req)
-	require.NoErrorf(h, err, "unable to send to route for %s", hn.Name())
-
-	return resp
-}
-
-// SendToRouteV2 makes a RPC call to SendToRouteV2 and asserts.
-func (h *HarnessTest) SendToRouteV2(hn *HarnessNode,
-	req *routerrpc.SendToRouteRequest) *lnrpc.HTLCAttempt {
-
-	ctxt, cancel := context.WithTimeout(h.runCtx, DefaultTimeout)
-	defer cancel()
-
-	resp, err := hn.rpc.Router.SendToRouteV2(ctxt, req)
-	require.NoErrorf(h, err, "unable to send to route v2 for %s", hn.Name())
-
-	return resp
-}
-
-// UpdateChannelPolicy makes a RPC call to UpdateChannelPolicy and asserts.
-func (h *HarnessTest) UpdateChannelPolicy(hn *HarnessNode,
-	req *lnrpc.PolicyUpdateRequest) *lnrpc.PolicyUpdateResponse {
-
-	ctxt, cancel := context.WithTimeout(h.runCtx, DefaultTimeout)
-	defer cancel()
-
-	resp, err := hn.rpc.LN.UpdateChannelPolicy(ctxt, req)
-	require.NoError(h, err, "failed to update policy")
-
-	return resp
 }
 
 // CleanupForceClose mines a force close commitment found in the mempool and
@@ -2321,245 +973,6 @@ func (h *HarnessTest) CleanupForceClose(hn *HarnessNode,
 	// The node should now sweep the funds, clean up by mining the sweeping
 	// tx.
 	h.MineBlocksAndAssertTx(1, 1)
-}
-
-// WaitForChannelPendingForceClose waits for the node to report that the
-// channel is pending force close, and that the UTXO nursery is aware of it.
-func (h *HarnessTest) WaitForChannelPendingForceClose(hn *HarnessNode,
-	fundingChanPoint *lnrpc.ChannelPoint) {
-
-	txid, err := lnrpc.GetChanPointFundingTxid(fundingChanPoint)
-	require.NoError(h, err, "failed to get chan point from txid")
-
-	op := wire.OutPoint{
-		Hash:  *txid,
-		Index: fundingChanPoint.OutputIndex,
-	}
-
-	err = wait.NoError(func() error {
-		pendingChanResp := h.GetPendingChannels(hn)
-
-		forceClose, err := FindForceClosedChannel(pendingChanResp, &op)
-		if err != nil {
-			return err
-		}
-
-		// We must wait until the UTXO nursery has received the channel
-		// and is aware of its maturity height.
-		if forceClose.MaturityHeight == 0 {
-			return fmt.Errorf("channel had maturity height of 0")
-		}
-
-		return nil
-	}, DefaultTimeout)
-
-	require.NoError(h, err, "timeout while finding force closed channels")
-}
-
-// AssertAmountPaid checks that the ListChannels command of the provided
-// node list the total amount sent and received as expected for the
-// provided channel.
-func (h *HarnessTest) AssertAmountPaid(channelName string, hn *HarnessNode,
-	chanPoint *lnrpc.ChannelPoint, amountSent, amountReceived int64) {
-
-	checkAmountPaid := func() error {
-		// Find the targeted channel.
-		channel, err := h.findChannel(hn, chanPoint)
-		if err != nil {
-			return fmt.Errorf("assert amount failed: %v", err)
-		}
-
-		if channel.TotalSatoshisSent != amountSent {
-			return fmt.Errorf("%v: incorrect amount"+
-				" sent: %v != %v", channelName,
-				channel.TotalSatoshisSent,
-				amountSent)
-		}
-		if channel.TotalSatoshisReceived !=
-			amountReceived {
-
-			return fmt.Errorf("%v: incorrect amount"+
-				" received: %v != %v",
-				channelName,
-				channel.TotalSatoshisReceived,
-				amountReceived)
-		}
-
-		return nil
-	}
-
-	// As far as HTLC inclusion in commitment transaction might be
-	// postponed we will try to check the balance couple of times,
-	// and then if after some period of time we receive wrong
-	// balance return the error.
-	err := wait.NoError(checkAmountPaid, DefaultTimeout)
-	require.NoError(h, err, "timeout while checking amount paid")
-}
-
-// WaitForNodeBlockHeight queries the node for its current block height until
-// it reaches the passed height.
-func (h *HarnessTest) WaitForNodeBlockHeight(hn *HarnessNode, height int32) {
-	err := wait.NoError(func() error {
-		info := h.GetInfo(hn)
-		if int32(info.BlockHeight) != height {
-			return fmt.Errorf("expected block height to "+
-				"be %v, was %v", height, info.BlockHeight)
-		}
-		return nil
-	}, DefaultTimeout)
-
-	require.NoError(h, err, "timeout while waiting for height")
-}
-
-// DescribeGraph makes a RPC call to the node's DescribeGraph and asserts. It
-// takes a bool to indicate whether we want to include private edges or not.
-func (h *HarnessTest) DescribeGraph(hn *HarnessNode,
-	includeUnannounced bool) *lnrpc.ChannelGraph {
-
-	ctxt, cancel := context.WithTimeout(h.runCtx, DefaultTimeout)
-	defer cancel()
-
-	req := &lnrpc.ChannelGraphRequest{
-		IncludeUnannounced: includeUnannounced,
-	}
-	resp, err := hn.rpc.LN.DescribeGraph(ctxt, req)
-	require.NoError(h, err, "failed to describe graph")
-
-	return resp
-}
-
-// AssertNumEdges checks that an expected number of edges can be found in the
-// node specified.
-func (h *HarnessTest) AssertNumEdges(hn *HarnessNode,
-	expected int, includeUnannounced bool) {
-
-	err := wait.NoError(func() error {
-		chanGraph := h.DescribeGraph(hn, includeUnannounced)
-
-		if len(chanGraph.Edges) == expected {
-			return nil
-		}
-
-		return fmt.Errorf("expected to find %d edge in the graph, "+
-			"found %d", expected, len(chanGraph.Edges))
-	}, DefaultTimeout)
-
-	require.NoError(h, err, "timeout while checking for edges")
-}
-
-// SubscribeChannelEvents creates a subscription client for channel events and
-// asserts its creation.
-func (h *HarnessTest) SubscribeChannelEvents(
-	hn *HarnessNode) lnrpc.Lightning_SubscribeChannelEventsClient {
-
-	req := &lnrpc.ChannelEventSubscription{}
-
-	// SubscribeChannelEvents needs to have the context alive for the
-	// entire test case as the returned client will be used for send and
-	// receive events stream. Thus we use runCtx here instead of a timeout
-	// context.
-	client, err := hn.rpc.LN.SubscribeChannelEvents(h.runCtx, req)
-	require.NoError(h, err, "unable to create channel update client")
-
-	return client
-}
-
-// LookupInvoice queries the node's invoices using the specified rHash.
-func (h *HarnessTest) LookupInvoice(hn *HarnessNode,
-	rHash []byte) *lnrpc.Invoice {
-
-	ctxt, cancel := context.WithTimeout(h.runCtx, DefaultTimeout)
-	defer cancel()
-
-	payHash := &lnrpc.PaymentHash{RHash: rHash}
-	resp, err := hn.rpc.LN.LookupInvoice(ctxt, payHash)
-	require.NoError(h, err, "unable to lookup invoice")
-
-	return resp
-}
-
-// LookupInvoiceV2 queries the node's invoices using the invoice client's
-// LookupInvoiceV2.
-func (h *HarnessTest) LookupInvoiceV2(hn *HarnessNode,
-	req *invoicesrpc.LookupInvoiceMsg) *lnrpc.Invoice {
-
-	ctxt, cancel := context.WithTimeout(h.runCtx, DefaultTimeout)
-	defer cancel()
-
-	resp, err := hn.rpc.Invoice.LookupInvoiceV2(ctxt, req)
-	require.NoError(h, err, "unable to lookup invoice")
-
-	return resp
-}
-
-// AssertLastHTLCError checks that the last sent HTLC of the last payment sent
-// by the given node failed with the expected failure code.
-func (h *HarnessTest) AssertLastHTLCError(hn *HarnessNode,
-	code lnrpc.Failure_FailureCode) {
-
-	// Use -1 to specify the last HTLC.
-	h.assertHTLCError(hn, code, -1)
-}
-
-// AssertFirstHTLCError checks that the first HTLC of the last payment sent
-// by the given node failed with the expected failure code.
-func (h *HarnessTest) AssertFirstHTLCError(hn *HarnessNode,
-	code lnrpc.Failure_FailureCode) {
-
-	// Use 0 to specify the first HTLC.
-	h.assertHTLCError(hn, code, 0)
-}
-
-// assertLastHTLCError checks that the HTLC at the specified index of the last
-// payment sent by the given node failed with the expected failure code.
-func (h *HarnessTest) assertHTLCError(hn *HarnessNode,
-	code lnrpc.Failure_FailureCode, index int) {
-
-	err := wait.NoError(func() error {
-		paymentsResp := h.ListPayments(hn, true)
-
-		payments := paymentsResp.Payments
-		if len(payments) == 0 {
-			return fmt.Errorf("no payments found")
-		}
-
-		payment := payments[len(payments)-1]
-		htlcs := payment.Htlcs
-		if len(htlcs) == 0 {
-			return fmt.Errorf("no htlcs found")
-		}
-
-		// If the index is greater than 0, check we have enough htlcs.
-		if index > 0 && len(htlcs) <= index {
-			return fmt.Errorf("not enough htlcs")
-		}
-
-		// If index is less than or equal to 0, we will read the last
-		// htlc.
-		if index <= 0 {
-			index = len(htlcs) - 1
-		}
-
-		htlc := htlcs[index]
-
-		// The htlc must have a status of failed.
-		if htlc.Status != lnrpc.HTLCAttempt_FAILED {
-			return fmt.Errorf("htlc should be failed")
-		}
-		// The failure field must not be empty.
-		if htlc.Failure == nil {
-			return fmt.Errorf("expected htlc failure")
-		}
-
-		// Exit if the expected code is found.
-		if htlc.Failure.Code == code {
-			return nil
-		}
-
-		return fmt.Errorf("unexpected failure code")
-	}, DefaultTimeout)
-
-	require.NoError(h, err, "timeout checking HTLC error")
 }
 
 // GetChanPointFundingTxid takes a channel point and converts it into a chain
@@ -2590,16 +1003,6 @@ func (h *HarnessTest) findChannel(hn *HarnessNode,
 	}
 
 	return nil, fmt.Errorf("channel not found")
-}
-
-// QueryChannelByChanPoint tries to find a channel matching the channel point
-// and asserts. It returns the channel found.
-func (h *HarnessTest) QueryChannelByChanPoint(hn *HarnessNode,
-	chanPoint *lnrpc.ChannelPoint) *lnrpc.Channel {
-
-	channel, err := h.findChannel(hn, chanPoint)
-	require.NoError(h, err, "failed to query channel")
-	return channel
 }
 
 // CreatePayReqs is a helper method that will create a slice of payment
@@ -2706,84 +1109,6 @@ func (h *HarnessTest) CompletePaymentRequests(hn *HarnessNode,
 	require.NoError(h, err, "timeout while checking for channel updates")
 }
 
-// AssertLocalBalance checks the local balance of the given channel is
-// expected. The channel found using the specified channel point is returned.
-func (h *HarnessTest) AssertLocalBalance(hn *HarnessNode,
-	cp *lnrpc.ChannelPoint, balance int64) *lnrpc.Channel {
-
-	var result *lnrpc.Channel
-
-	// Get the funding point.
-	err := wait.NoError(func() error {
-		// Find the target channel first.
-		target, err := h.findChannel(hn, cp)
-
-		// Exit early if the channel is not found.
-		if err != nil {
-			return fmt.Errorf("check balance failed: %v", err)
-		}
-
-		result = target
-
-		// Check local balance.
-		if target.LocalBalance == balance {
-			return nil
-		}
-
-		return fmt.Errorf("balance is incorrect, got %v, expected %v",
-			target.LocalBalance, balance)
-	}, DefaultTimeout)
-
-	require.NoError(h, err, "timeout while chekcing for balance")
-
-	return result
-}
-
-// AssertChannelState asserts the channel state by checking the values in
-// fields, LocalBalance, RemoteBalance and num of PendingHtlcs.
-func (h *HarnessTest) AssertChannelState(hn *HarnessNode,
-	cp *lnrpc.ChannelPoint, localBalance,
-	remoteBalance int64, numPendingHtlcs int) *lnrpc.Channel {
-
-	var result *lnrpc.Channel
-
-	// Get the funding point.
-	err := wait.NoError(func() error {
-		// Find the target channel first.
-		target, err := h.findChannel(hn, cp)
-
-		// Exit early if the channel is not found.
-		if err != nil {
-			return fmt.Errorf("check balance failed: %v", err)
-		}
-
-		if len(target.PendingHtlcs) != numPendingHtlcs {
-			return fmt.Errorf("pending htlcs is "+
-				"incorrect, got %v, expected %v",
-				len(target.PendingHtlcs), 0)
-		}
-
-		if target.LocalBalance != localBalance {
-			return fmt.Errorf("local balance is "+
-				"incorrect, got %v, expected %v",
-				target.LocalBalance, localBalance)
-		}
-
-		if target.RemoteBalance != remoteBalance {
-			return fmt.Errorf("remote balance is "+
-				"incorrect, got %v, expected %v",
-				target.RemoteBalance, remoteBalance)
-		}
-
-		result = target
-		return nil
-	}, DefaultTimeout)
-
-	require.NoError(h, err, "timeout while chekcing for balance")
-
-	return result
-}
-
 // BackupDb created a db backup for the specified node and asserts.
 func (h *HarnessTest) BackupDb(hn *HarnessNode) {
 	require.NoError(h, h.net.BackupDb(hn), "failed to copy db files")
@@ -2792,1101 +1117,6 @@ func (h *HarnessTest) BackupDb(hn *HarnessNode) {
 // RestoreDb restores a db backup for the specified node and asserts.
 func (h *HarnessTest) RestoreDb(hn *HarnessNode) {
 	require.NoError(h, h.net.RestoreDb(hn), "failed to restore db")
-}
-
-// AssertNumUpdates checks the num of updates is expected from the given
-// channel.
-//
-// NOTE: doesn't wait.
-func (h *HarnessTest) AssertNumUpdates(hn *HarnessNode,
-	num uint64, cp *lnrpc.ChannelPoint) {
-
-	// Find the target channel first.
-	target, err := h.findChannel(hn, cp)
-	require.NoError(h, err, "unable to find channel")
-
-	require.Equal(h, num, target.NumUpdates, "NumUpdates not matched")
-}
-
-// ExportAllChanBackups makes a RPC call to the node's ExportAllChannelBackups
-// and asserts.
-func (h *HarnessTest) ExportAllChanBackups(
-	hn *HarnessNode) *lnrpc.ChanBackupSnapshot {
-
-	ctxt, cancel := context.WithTimeout(h.runCtx, DefaultTimeout)
-	defer cancel()
-
-	req := &lnrpc.ChanBackupExportRequest{}
-	chanBackup, err := hn.rpc.LN.ExportAllChannelBackups(ctxt, req)
-	require.NoError(h, err, "unable to obtain channel backup")
-
-	return chanBackup
-}
-
-// ExportChanBackup makes a RPC call to the node's ExportChannelBackup
-// and asserts.
-func (h *HarnessTest) ExportChanBackup(hn *HarnessNode,
-	chanPoint *lnrpc.ChannelPoint) *lnrpc.ChannelBackup {
-
-	ctxt, cancel := context.WithTimeout(h.runCtx, DefaultTimeout)
-	defer cancel()
-
-	req := &lnrpc.ExportChannelBackupRequest{
-		ChanPoint: chanPoint,
-	}
-	chanBackup, err := hn.rpc.LN.ExportChannelBackup(ctxt, req)
-	require.NoError(h, err, "unable to obtain channel backup")
-
-	return chanBackup
-}
-
-// RestoreChanBackups makes a RPC call to the node's RestoreChannelBackups and
-// asserts.
-func (h *HarnessTest) RestoreChanBackups(hn *HarnessNode,
-	req *lnrpc.RestoreChanBackupRequest) *lnrpc.RestoreBackupResponse {
-
-	ctxt, cancel := context.WithTimeout(h.runCtx, DefaultTimeout)
-	defer cancel()
-
-	resp, err := hn.rpc.LN.RestoreChannelBackups(ctxt, req)
-	require.NoError(h, err, "unable to restore backups")
-
-	return resp
-}
-
-// FundPsbt makes a RPC call to node's FundPsbt and asserts.
-func (h *HarnessTest) FundPsbt(hn *HarnessNode,
-	req *walletrpc.FundPsbtRequest) *walletrpc.FundPsbtResponse {
-
-	ctxt, cancel := context.WithTimeout(h.runCtx, DefaultTimeout)
-	defer cancel()
-
-	resp, err := hn.rpc.WalletKit.FundPsbt(ctxt, req)
-	require.NoError(h, err, "fund psbt failed")
-
-	return resp
-}
-
-// FinalizePsbt makes a RPC call to node's FundPsbt and asserts.
-func (h *HarnessTest) FinalizePsbt(hn *HarnessNode,
-	req *walletrpc.FinalizePsbtRequest) *walletrpc.FinalizePsbtResponse {
-
-	ctxt, cancel := context.WithTimeout(h.runCtx, DefaultTimeout)
-	defer cancel()
-
-	resp, err := hn.rpc.WalletKit.FinalizePsbt(ctxt, req)
-	require.NoError(h, err, "finalize psbt failed")
-
-	return resp
-}
-
-// AssertNodeStarted waits until the node is fully started or asserts a timeout
-// error.
-func (h *HarnessTest) AssertNodeStarted(hn *HarnessNode) {
-	require.NoError(h, hn.WaitUntilServerActive())
-}
-
-// AssertNumUTXOs waits for the given number of UTXOs to be available or fails
-// if that isn't the case before the default timeout.
-func (h *HarnessTest) AssertNumUTXOs(hn *HarnessNode, expectedUtxos int) {
-	err := wait.NoError(func() error {
-		resp := h.ListUnspent(hn, "", math.MaxInt32, 1)
-		if len(resp.Utxos) != expectedUtxos {
-			return fmt.Errorf("not enough UTXOs, got %d wanted %d",
-				len(resp.Utxos), expectedUtxos)
-		}
-
-		return nil
-	}, DefaultTimeout)
-	require.NoError(h, err, "timeout waiting for UTXOs")
-}
-
-type BackupSubscriber lnrpc.Lightning_SubscribeChannelBackupsClient
-
-// SubscribeChannelBackups creates a client to listen to channel backup stream.
-func (h *HarnessTest) SubscribeChannelBackups(
-	hn *HarnessNode) BackupSubscriber {
-
-	// Use runCtx here instead of timeout context to keep the stream client
-	// alive.
-	backupStream, err := hn.rpc.LN.SubscribeChannelBackups(
-		h.runCtx, &lnrpc.ChannelBackupSubscription{},
-	)
-	require.NoError(h, err, "unable to create backup stream")
-
-	return backupStream
-}
-
-// VerifyChanBackup makes a RPC call to node's VerifyChanBackup and asserts.
-func (h *HarnessTest) VerifyChanBackup(hn *HarnessNode,
-	ss *lnrpc.ChanBackupSnapshot) *lnrpc.VerifyChanBackupResponse {
-
-	ctxt, cancel := context.WithTimeout(h.runCtx, DefaultTimeout)
-	defer cancel()
-
-	resp, err := hn.rpc.LN.VerifyChanBackup(ctxt, ss)
-	require.NoError(h, err, "unable to verify backup")
-
-	return resp
-}
-
-// SignMessage makes a RPC call to node's SignMessage and asserts.
-func (h *HarnessTest) SignMessage(hn *HarnessNode,
-	msg []byte) *lnrpc.SignMessageResponse {
-
-	ctxt, cancel := context.WithTimeout(h.runCtx, DefaultTimeout)
-	defer cancel()
-
-	req := &lnrpc.SignMessageRequest{Msg: msg}
-	resp, err := hn.rpc.LN.SignMessage(ctxt, req)
-	require.NoError(h, err, "SignMessage rpc call failed")
-
-	return resp
-}
-
-// VerifyMessage makes a RPC call to node's VerifyMessage and asserts.
-func (h *HarnessTest) VerifyMessage(hn *HarnessNode, msg []byte,
-	sig string) *lnrpc.VerifyMessageResponse {
-
-	ctxt, cancel := context.WithTimeout(h.runCtx, DefaultTimeout)
-	defer cancel()
-
-	req := &lnrpc.VerifyMessageRequest{Msg: msg, Signature: sig}
-	resp, err := hn.rpc.LN.VerifyMessage(ctxt, req)
-	require.NoError(h, err, "VerifyMessage failed")
-
-	return resp
-}
-
-// ClosedChannels makes a RPC call to node's ClosedChannels and asserts.
-func (h *HarnessTest) ClosedChannels(hn *HarnessNode,
-	abandoned bool) *lnrpc.ClosedChannelsResponse {
-
-	ctxt, cancel := context.WithTimeout(h.runCtx, DefaultTimeout)
-	defer cancel()
-
-	req := &lnrpc.ClosedChannelsRequest{
-		Abandoned: abandoned,
-	}
-	resp, err := hn.rpc.LN.ClosedChannels(ctxt, req)
-	require.NoError(h, err, "list closed channels failed")
-
-	return resp
-}
-
-// AssertZombieChannel asserts that a given channel found using the chanID is
-// marked as zombie.
-func (h *HarnessTest) AssertZombieChannel(hn *HarnessNode, chanID uint64) {
-	ctxt, cancel := context.WithTimeout(h.runCtx, DefaultTimeout)
-	defer cancel()
-
-	err := wait.NoError(func() error {
-		_, err := hn.rpc.LN.GetChanInfo(
-			ctxt, &lnrpc.ChanInfoRequest{ChanId: chanID},
-		)
-		if err == nil {
-			return fmt.Errorf("expected error but got nil")
-		}
-
-		if !strings.Contains(err.Error(), "marked as zombie") {
-			return fmt.Errorf("expected error to contain '%s' but "+
-				"was '%v'", "marked as zombie", err)
-		}
-
-		return nil
-	}, DefaultTimeout)
-	require.NoError(h, err, "timeout while checking zombie channel")
-}
-
-// LabelTransactionAssertErr makes a RPC call to the node's LabelTransaction
-// and asserts an error is returned. It then returns the error.
-func (h *HarnessTest) LabelTransactionAssertErr(hn *HarnessNode,
-	req *walletrpc.LabelTransactionRequest) error {
-
-	ctxt, cancel := context.WithTimeout(h.runCtx, DefaultTimeout)
-	defer cancel()
-
-	_, err := hn.rpc.WalletKit.LabelTransaction(ctxt, req)
-	require.Error(h, err, "expected error returned")
-
-	return err
-}
-
-// LabelTransaction makes a RPC call to the node's LabelTransaction
-// and asserts no error is returned.
-func (h *HarnessTest) LabelTransaction(hn *HarnessNode,
-	req *walletrpc.LabelTransactionRequest) {
-
-	ctxt, cancel := context.WithTimeout(h.runCtx, DefaultTimeout)
-	defer cancel()
-
-	_, err := hn.rpc.WalletKit.LabelTransaction(ctxt, req)
-	require.NoError(h, err, "failed to label transaction")
-}
-
-// UpdateChanStatus makes a UpdateChanStatus RPC call to node's RouterClient
-// and asserts.
-func (h *HarnessTest) UpdateChanStatus(hn *HarnessNode,
-	req *routerrpc.UpdateChanStatusRequest) *routerrpc.UpdateChanStatusResponse {
-
-	ctxt, cancel := context.WithTimeout(h.runCtx, DefaultTimeout)
-	defer cancel()
-
-	resp, err := hn.rpc.Router.UpdateChanStatus(ctxt, req)
-	require.NoErrorf(h, err, "UpdateChanStatus failed")
-
-	return resp
-}
-
-// AssertNumChannelUpdates asserts that a given number of channel updates has
-// been seen in the specified node.
-func (h *HarnessTest) AssertNumChannelUpdates(hn *HarnessNode,
-	chanPoint *lnrpc.ChannelPoint, num int) {
-
-	op := h.OutPointFromChannelPoint(chanPoint)
-	err := hn.WaitForNetworkNumChanUpdates(op, num)
-	require.NoError(h, err, "failed to assert num of channel updates")
-}
-
-// AssertNumNodeAnns asserts that a given number of node announcements has been
-// seen in the specified node.
-func (h *HarnessTest) AssertNumNodeAnns(hn *HarnessNode,
-	pubkey string, num int) {
-
-	// We will get the current number of channel updates first and add it
-	// to our expected number of newly created channel updates.
-	err := hn.WaitForNetworkNumNodeUpdates(pubkey, num)
-	require.NoError(h, err, "failed to assert num of channel updates")
-}
-
-// AssertChannelClosedFromGraph asserts a given channel is closed by checking
-// the graph topology subscription of the specified node. Returns the closed
-// channel update if found.
-func (h *HarnessTest) AssertChannelClosedFromGraph(hn *HarnessNode,
-	chanPoint *lnrpc.ChannelPoint) *lnrpc.ClosedChannelUpdate {
-
-	closedChan, err := hn.WaitForNetworkChannelClose(chanPoint)
-	require.NoError(h, err, "failed to wait for channel close")
-
-	return closedChan
-}
-
-type (
-	AnnReq  *peersrpc.NodeAnnouncementUpdateRequest
-	AnnResp *peersrpc.NodeAnnouncementUpdateResponse
-)
-
-// UpdateNodeAnnouncement makes an UpdateNodeAnnouncement RPC call the the
-// peersrpc client and asserts.
-func (h *HarnessTest) UpdateNodeAnnouncement(hn *HarnessNode,
-	req AnnReq) AnnResp {
-
-	ctxt, cancel := context.WithTimeout(h.runCtx, DefaultTimeout)
-	defer cancel()
-
-	resp, err := hn.rpc.Peer.UpdateNodeAnnouncement(ctxt, req)
-	require.NoError(h, err, "failed to update announcement")
-
-	return resp
-}
-
-// UpdateNodeAnnouncementErr makes an UpdateNodeAnnouncement RPC call the the
-// peersrpc client and asserts an error is returned.
-func (h *HarnessTest) UpdateNodeAnnouncementErr(hn *HarnessNode, req AnnReq) {
-	ctxt, cancel := context.WithTimeout(h.runCtx, DefaultTimeout)
-	defer cancel()
-
-	_, err := hn.rpc.Peer.UpdateNodeAnnouncement(ctxt, req)
-	require.Error(h, err, "expect an error from update announcement")
-}
-
-// ListSweeps makes a ListSweeps RPC call to the node's WalletKit client.
-func (h *HarnessTest) ListSweeps(hn *HarnessNode,
-	verbose bool) *walletrpc.ListSweepsResponse {
-
-	ctxt, cancel := context.WithTimeout(h.runCtx, DefaultTimeout)
-	defer cancel()
-
-	req := &walletrpc.ListSweepsRequest{Verbose: verbose}
-	resp, err := hn.rpc.WalletKit.ListSweeps(ctxt, req)
-	require.NoError(h, err, "failed to ListSweeps")
-
-	return resp
-}
-
-// ConnectPeerAssertErr makes a RPC call to ConnectPeer and asserts an error
-// returned.
-func (h *HarnessTest) ConnectPeerAssertErr(hn *HarnessNode,
-	req *lnrpc.ConnectPeerRequest) error {
-
-	ctxt, cancel := context.WithTimeout(h.runCtx, DefaultTimeout)
-	defer cancel()
-
-	_, err := hn.rpc.LN.ConnectPeer(ctxt, req)
-	require.Error(h, err, "expected an error from ConnectPeer")
-
-	return err
-}
-
-// DecodePayReq makes a RPC call to node's DecodePayReq and asserts.
-func (h *HarnessTest) DecodePayReq(hn *HarnessNode, req string) *lnrpc.PayReq {
-	ctxt, cancel := context.WithTimeout(h.runCtx, DefaultTimeout)
-	defer cancel()
-
-	payReq := &lnrpc.PayReqString{PayReq: req}
-	resp, err := hn.rpc.LN.DecodePayReq(ctxt, payReq)
-	require.NoError(h, err, "failed to decode pay req")
-
-	return resp
-}
-
-type HtlcEventsClient routerrpc.Router_SubscribeHtlcEventsClient
-
-// SubscribeHtlcEvents makes a subscription to the HTLC events and returns a
-// htlc event client.
-func (h *HarnessTest) SubscribeHtlcEvents(hn *HarnessNode) HtlcEventsClient {
-	// Use runCtx here to keep the client alive for the scope of the test.
-	client, err := hn.rpc.Router.SubscribeHtlcEvents(
-		h.runCtx, &routerrpc.SubscribeHtlcEventsRequest{},
-	)
-	require.NoError(h, err, "could not subscribe events")
-
-	return client
-}
-
-// ReceiveHtlcEvent waits until a message is received on the subscribe
-// htlc event stream or the timeout is reached.
-func (h *HarnessTest) ReceiveHtlcEvent(
-	stream HtlcEventsClient) *routerrpc.HtlcEvent {
-
-	chanMsg := make(chan *routerrpc.HtlcEvent)
-	errChan := make(chan error)
-	go func() {
-		// Consume one message. This will block until the message is
-		// received.
-		resp, err := stream.Recv()
-		if err != nil {
-			errChan <- err
-			return
-		}
-		chanMsg <- resp
-	}()
-
-	select {
-	case <-time.After(DefaultTimeout):
-		require.Fail(h, "timeout", "timeout receiving htlc "+
-			"event update")
-
-	case err := <-errChan:
-		require.Failf(h, "err from stream",
-			"received err from stream: %v", err)
-
-	case updateMsg := <-chanMsg:
-		return updateMsg
-	}
-
-	return nil
-}
-
-// AssertHtlcEvents consumes events from a client and ensures that they are of
-// the expected type and contain the expected number of forwards, forward
-// failures and settles.
-func (h *HarnessTest) AssertHtlcEvents(client HtlcEventsClient,
-	fwdCount, fwdFailCount, settleCount, linkFailCount int,
-	userType routerrpc.HtlcEvent_EventType) []*routerrpc.HtlcEvent {
-
-	var forwards, forwardFails, settles, linkFails int
-
-	numEvents := fwdCount + fwdFailCount + settleCount + linkFailCount
-	events := make([]*routerrpc.HtlcEvent, 0)
-
-	for i := 0; i < numEvents; i++ {
-		event := h.ReceiveHtlcEvent(client)
-		require.Equalf(h, userType, event.EventType,
-			"wrong event type, want %v got %v",
-			userType, event.EventType)
-
-		events = append(events, event)
-
-		switch event.Event.(type) {
-		case *routerrpc.HtlcEvent_ForwardEvent:
-			forwards++
-
-		case *routerrpc.HtlcEvent_ForwardFailEvent:
-			forwardFails++
-
-		case *routerrpc.HtlcEvent_SettleEvent:
-			settles++
-
-		case *routerrpc.HtlcEvent_LinkFailEvent:
-			linkFails++
-
-		default:
-			require.Fail(h, "assert event fail",
-				"unexpected event: %T", event.Event)
-		}
-	}
-
-	require.Equal(h, fwdCount, forwards, "num of forwards mismatch")
-	require.Equal(h, fwdFailCount, forwardFails,
-		"num of forward fails mismatch")
-	require.Equal(h, settleCount, settles, "num of settles mismatch")
-	require.Equal(h, linkFailCount, linkFails, "num of link fails mismatch")
-
-	return events
-}
-
-// AssertFeeReport checks that the fee report from the given node has the
-// desired day, week, and month sum values.
-func (h *HarnessTest) AssertFeeReport(hn *HarnessNode,
-	day, week, month int) {
-
-	ctxt, cancel := context.WithTimeout(h.runCtx, DefaultTimeout)
-	defer cancel()
-
-	feeReport, err := hn.rpc.LN.FeeReport(ctxt, &lnrpc.FeeReportRequest{})
-	require.NoError(h, err, "unable to query for fee report")
-
-	require.EqualValues(h, day, feeReport.DayFeeSum, "day fee mismatch")
-	require.EqualValues(h, week, feeReport.DayFeeSum, "day week mismatch")
-	require.EqualValues(h, month, feeReport.DayFeeSum, "day month mismatch")
-}
-
-// ForwardingHistory makes a RPC call to the node's ForwardingHistory and
-// asserts.
-func (h *HarnessTest) ForwardingHistory(
-	hn *HarnessNode) *lnrpc.ForwardingHistoryResponse {
-
-	ctxt, cancel := context.WithTimeout(h.runCtx, DefaultTimeout)
-	defer cancel()
-
-	resp, err := hn.rpc.LN.ForwardingHistory(
-		ctxt, &lnrpc.ForwardingHistoryRequest{},
-	)
-	require.NoError(h, err, "failed to query forwarding history")
-
-	return resp
-}
-
-// DeleteAllPayments makes a RPC call to the node's DeleteAllPayments and
-// asserts.
-func (h *HarnessTest) DeleteAllPayments(hn *HarnessNode) {
-	ctxt, cancel := context.WithTimeout(h.runCtx, DefaultTimeout)
-	defer cancel()
-
-	req := &lnrpc.DeleteAllPaymentsRequest{}
-	_, err := hn.rpc.LN.DeleteAllPayments(ctxt, req)
-	require.NoError(h, err, "unable to delete payments")
-}
-
-// GetMissionControlConfig makes a RPC call to the node's
-// GetMissionControlConfig and asserts.
-func (h *HarnessTest) GetMissionControlConfig(
-	hn *HarnessNode) *routerrpc.GetMissionControlConfigResponse {
-
-	ctxt, cancel := context.WithTimeout(h.runCtx, DefaultTimeout)
-	defer cancel()
-
-	req := &routerrpc.GetMissionControlConfigRequest{}
-	resp, err := hn.rpc.Router.GetMissionControlConfig(ctxt, req)
-	require.NoError(h, err, "get mission control failed")
-
-	return resp
-}
-
-// SetMissionControlConfig makes a RPC call to the node's
-// SetMissionControlConfig and asserts.
-func (h *HarnessTest) SetMissionControlConfig(hn *HarnessNode,
-	config *routerrpc.MissionControlConfig) {
-
-	ctxt, cancel := context.WithTimeout(h.runCtx, DefaultTimeout)
-	defer cancel()
-
-	req := &routerrpc.SetMissionControlConfigRequest{Config: config}
-	_, err := hn.rpc.Router.SetMissionControlConfig(ctxt, req)
-	require.NoError(h, err, "set mission control failed")
-}
-
-// ResetMissionControl makes a RPC call to the node's ResetMissionControl and
-// asserts.
-func (h *HarnessTest) ResetMissionControl(hn *HarnessNode) {
-	ctxt, cancel := context.WithTimeout(h.runCtx, DefaultTimeout)
-	defer cancel()
-
-	req := &routerrpc.ResetMissionControlRequest{}
-	_, err := hn.rpc.Router.ResetMissionControl(ctxt, req)
-	require.NoError(h, err, "reset mission control failed")
-}
-
-// QueryProbability makes a RPC call to the node's QueryProbability and
-// asserts.
-func (h *HarnessTest) QueryProbability(hn *HarnessNode,
-	req *routerrpc.QueryProbabilityRequest) *routerrpc.QueryProbabilityResponse {
-
-	ctxt, cancel := context.WithTimeout(h.runCtx, DefaultTimeout)
-	defer cancel()
-
-	resp, err := hn.rpc.Router.QueryProbability(ctxt, req)
-	require.NoError(h, err, "query probability failed")
-
-	return resp
-}
-
-// XImportMissionControl makes a RPC call to the node's XImportMissionControl
-// and asserts.
-func (h *HarnessTest) XImportMissionControl(hn *HarnessNode,
-	req *routerrpc.XImportMissionControlRequest) {
-
-	ctxt, cancel := context.WithTimeout(h.runCtx, DefaultTimeout)
-	defer cancel()
-
-	_, err := hn.rpc.Router.XImportMissionControl(ctxt, req)
-	require.NoError(h, err, "x import mission control failed")
-}
-
-// XImportMissionControlAssertErr makes a RPC call to the node's XImportMissionControl
-// and asserts an error occurred.
-func (h *HarnessTest) XImportMissionControlAssertErr(hn *HarnessNode,
-	req *routerrpc.XImportMissionControlRequest) {
-
-	ctxt, cancel := context.WithTimeout(h.runCtx, DefaultTimeout)
-	defer cancel()
-
-	_, err := hn.rpc.Router.XImportMissionControl(ctxt, req)
-	require.Error(h, err, "expect an error from x import mission control")
-}
-
-type InvoiceUpdateClient lnrpc.Lightning_SubscribeInvoicesClient
-
-// SubscribeInvoices creates a subscription client for invoice events and
-// asserts its creation.
-func (h *HarnessTest) SubscribeInvoices(hn *HarnessNode,
-	req *lnrpc.InvoiceSubscription) InvoiceUpdateClient {
-
-	// SubscribeInvoices needs to have the context alive for the
-	// entire test case as the returned client will be used for send and
-	// receive events stream. Thus we use runCtx here instead of a timeout
-	// context.
-	client, err := hn.rpc.LN.SubscribeInvoices(h.runCtx, req)
-	require.NoError(h, err, "unable to create invoice subscription client")
-
-	return client
-}
-
-// ReceiveInvoiceUpdate waits until a message is received on the subscribe
-// invoice stream or the timeout is reached.
-func (h *HarnessTest) ReceiveInvoiceUpdate(
-	stream InvoiceUpdateClient) *lnrpc.Invoice {
-
-	chanMsg := make(chan *lnrpc.Invoice)
-	errChan := make(chan error)
-	go func() {
-		// Consume one message. This will block until the message is
-		// received.
-		resp, err := stream.Recv()
-		if err != nil {
-			errChan <- err
-			return
-		}
-		chanMsg <- resp
-	}()
-
-	select {
-	case <-time.After(DefaultTimeout):
-		require.Fail(h, "timeout", "timeout receiving invoice update")
-
-	case err := <-errChan:
-		require.Failf(h, "err from stream",
-			"received err from stream: %v", err)
-
-	case updateMsg := <-chanMsg:
-		return updateMsg
-	}
-
-	return nil
-}
-
-// GetInfoWatchtower makes a RPC call to the watchtower of the given node and
-// asserts.
-func (h *HarnessTest) GetInfoWatchtower(
-	hn *HarnessNode) *watchtowerrpc.GetInfoResponse {
-
-	ctxt, cancel := context.WithTimeout(h.runCtx, DefaultTimeout)
-	defer cancel()
-
-	req := &watchtowerrpc.GetInfoRequest{}
-	info, err := hn.rpc.Watchtower.GetInfo(ctxt, req)
-	require.NoError(h, err, "failed to GetInfo from watchtower")
-
-	return info
-}
-
-// AddTower makes a RPC call to the WatchtowerClient of the given node and
-// asserts.
-func (h *HarnessTest) AddTower(hn *HarnessNode,
-	req *wtclientrpc.AddTowerRequest) *wtclientrpc.AddTowerResponse {
-
-	ctxt, cancel := context.WithTimeout(h.runCtx, DefaultTimeout)
-	defer cancel()
-
-	resp, err := hn.rpc.WatchtowerClient.AddTower(ctxt, req)
-	require.NoError(h, err, "failed to AddTower")
-
-	return resp
-}
-
-// WatchtowerStats makes a RPC call to the WatchtowerClient of the given node
-// and asserts.
-func (h *HarnessTest) WatchtowerStats(
-	hn *HarnessNode) *wtclientrpc.StatsResponse {
-
-	ctxt, cancel := context.WithTimeout(h.runCtx, DefaultTimeout)
-	defer cancel()
-
-	req := &wtclientrpc.StatsRequest{}
-	resp, err := hn.rpc.WatchtowerClient.Stats(ctxt, req)
-	require.NoError(h, err, "failed to call Stats")
-
-	return resp
-}
-
-type TrackPaymentClient routerrpc.Router_TrackPaymentV2Client
-
-// TrackPaymentV2 creates a subscription client for given invoice and
-// asserts its creation.
-func (h *HarnessTest) TrackPaymentV2(hn *HarnessNode,
-	payHash []byte) TrackPaymentClient {
-
-	req := &routerrpc.TrackPaymentRequest{PaymentHash: payHash}
-
-	// TrackPaymentV2 needs to have the context alive for the entire test
-	// case as the returned client will be used for send and receive events
-	// stream. Thus we use runCtx here instead of a timeout context.
-	client, err := hn.rpc.Router.TrackPaymentV2(h.runCtx, req)
-	require.NoError(h, err, "unable to create track payment client")
-
-	return client
-}
-
-// ReceiveTrackPayment waits until a message is received on the track payment
-// stream or the timeout is reached.
-func (h *HarnessTest) ReceiveTrackPayment(
-	stream TrackPaymentClient) *lnrpc.Payment {
-
-	chanMsg := make(chan *lnrpc.Payment)
-	errChan := make(chan error)
-	go func() {
-		// Consume one message. This will block until the message is
-		// received.
-		resp, err := stream.Recv()
-		if err != nil {
-			errChan <- err
-			return
-		}
-		chanMsg <- resp
-	}()
-
-	select {
-	case <-time.After(DefaultTimeout):
-		require.Fail(h, "timeout", "timeout trakcing payment")
-
-	case err := <-errChan:
-		require.Failf(h, "err from stream",
-			"received err from stream: %v", err)
-
-	case updateMsg := <-chanMsg:
-		return updateMsg
-	}
-
-	return nil
-}
-
-// WaitForBlockchainSync waits until the node is synced to chain.
-func (h *HarnessTest) WaitForBlockchainSync(hn *HarnessNode) {
-	err := hn.WaitForBlockchainSync()
-	require.NoError(h, err, "timeout waiting for blockchain sync")
-}
-
-// DeriveSharedKey makes a RPC call to the node's SignerClient and asserts.
-func (h *HarnessTest) DeriveSharedKey(hn *HarnessNode,
-	req *signrpc.SharedKeyRequest) *signrpc.SharedKeyResponse {
-
-	ctxt, cancel := context.WithTimeout(h.runCtx, DefaultTimeout)
-	defer cancel()
-
-	resp, err := hn.rpc.Signer.DeriveSharedKey(ctxt, req)
-	require.NoError(h, err, "calling DeriveSharedKey failed")
-	return resp
-}
-
-// DeriveSharedKeyErr makes a RPC call to the node's SignerClient and asserts
-// there is an error.
-func (h *HarnessTest) DeriveSharedKeyErr(hn *HarnessNode,
-	req *signrpc.SharedKeyRequest) error {
-
-	ctxt, cancel := context.WithTimeout(h.runCtx, DefaultTimeout)
-	defer cancel()
-
-	_, err := hn.rpc.Signer.DeriveSharedKey(ctxt, req)
-	require.Error(h, err, "expected error from calling DeriveSharedKey")
-	return err
-}
-
-// SignOutputRaw makes a RPC call to the node's SignerClient and asserts.
-func (h *HarnessTest) SignOutputRaw(hn *HarnessNode,
-	req *signrpc.SignReq) *signrpc.SignResp {
-
-	ctxt, cancel := context.WithTimeout(h.runCtx, DefaultTimeout)
-	defer cancel()
-
-	resp, err := hn.rpc.Signer.SignOutputRaw(ctxt, req)
-	require.NoError(h, err, "failed to sign raw output")
-
-	return resp
-}
-
-// SignOutputRawErr makes a RPC call to the node's SignerClient and asserts an
-// error is returned.
-func (h *HarnessTest) SignOutputRawErr(hn *HarnessNode,
-	req *signrpc.SignReq) error {
-
-	ctxt, cancel := context.WithTimeout(h.runCtx, DefaultTimeout)
-	defer cancel()
-
-	_, err := hn.rpc.Signer.SignOutputRaw(ctxt, req)
-	require.Error(h, err, "expect to fail to sign raw output")
-
-	return err
-}
-
-// MuSig2CreateSession makes a RPC call to the node's SignerClient and asserts.
-func (h *HarnessTest) MuSig2CreateSession(hn *HarnessNode,
-	req *signrpc.MuSig2SessionRequest) *signrpc.MuSig2SessionResponse {
-
-	ctxt, cancel := context.WithTimeout(h.runCtx, DefaultTimeout)
-	defer cancel()
-
-	resp, err := hn.rpc.Signer.MuSig2CreateSession(ctxt, req)
-	require.NoError(h, err, "failed to create musig2 session")
-
-	return resp
-}
-
-// MuSig2CombineKeys makes a RPC call to the node's SignerClient and asserts.
-func (h *HarnessTest) MuSig2CombineKeys(hn *HarnessNode,
-	req *signrpc.MuSig2CombineKeysRequest) *signrpc.MuSig2CombineKeysResponse {
-
-	ctxt, cancel := context.WithTimeout(h.runCtx, DefaultTimeout)
-	defer cancel()
-
-	resp, err := hn.rpc.Signer.MuSig2CombineKeys(ctxt, req)
-	require.NoError(h, err, "failed to combine keys")
-
-	return resp
-}
-
-// MuSig2RegisterNonces makes a RPC call to the node's SignerClient and asserts.
-func (h *HarnessTest) MuSig2RegisterNonces(hn *HarnessNode,
-	req *signrpc.MuSig2RegisterNoncesRequest) *signrpc.MuSig2RegisterNoncesResponse {
-
-	ctxt, cancel := context.WithTimeout(h.runCtx, DefaultTimeout)
-	defer cancel()
-
-	resp, err := hn.rpc.Signer.MuSig2RegisterNonces(ctxt, req)
-	require.NoError(h, err, "failed to register nonces")
-
-	return resp
-}
-
-// MuSig2Sign makes a RPC call to the node's SignerClient and asserts.
-func (h *HarnessTest) MuSig2Sign(hn *HarnessNode,
-	req *signrpc.MuSig2SignRequest) *signrpc.MuSig2SignResponse {
-
-	ctxt, cancel := context.WithTimeout(h.runCtx, DefaultTimeout)
-	defer cancel()
-
-	resp, err := hn.rpc.Signer.MuSig2Sign(ctxt, req)
-	require.NoError(h, err, "failed to sign")
-
-	return resp
-}
-
-// MuSig2SignErr makes a RPC call to the node's SignerClient and asserts an
-// error is returned.
-func (h *HarnessTest) MuSig2SignErr(hn *HarnessNode,
-	req *signrpc.MuSig2SignRequest) error {
-
-	ctxt, cancel := context.WithTimeout(h.runCtx, DefaultTimeout)
-	defer cancel()
-
-	_, err := hn.rpc.Signer.MuSig2Sign(ctxt, req)
-	require.Error(h, err, "expect an error")
-
-	return err
-}
-
-// MuSig2CombineSig makes a RPC call to the node's SignerClient and asserts.
-func (h *HarnessTest) MuSig2CombineSig(hn *HarnessNode,
-	req *signrpc.MuSig2CombineSigRequest) *signrpc.MuSig2CombineSigResponse {
-
-	ctxt, cancel := context.WithTimeout(h.runCtx, DefaultTimeout)
-	defer cancel()
-
-	resp, err := hn.rpc.Signer.MuSig2CombineSig(ctxt, req)
-	require.NoError(h, err, "failed to combine sig")
-
-	return resp
-}
-
-// MuSig2Cleanup makes a RPC call to the node's SignerClient and asserts.
-func (h *HarnessTest) MuSig2Cleanup(hn *HarnessNode,
-	req *signrpc.MuSig2CleanupRequest) *signrpc.MuSig2CleanupResponse {
-
-	ctxt, cancel := context.WithTimeout(h.runCtx, DefaultTimeout)
-	defer cancel()
-
-	resp, err := hn.rpc.Signer.MuSig2Cleanup(ctxt, req)
-	require.NoError(h, err, "failed to cleanup")
-
-	return resp
-}
-
-// BumpFee makes a RPC call to the node's WalletKitClient and asserts.
-func (h *HarnessTest) BumpFee(hn *HarnessNode,
-	req *walletrpc.BumpFeeRequest) *walletrpc.BumpFeeResponse {
-
-	ctxt, cancel := context.WithTimeout(h.runCtx, DefaultTimeout)
-	defer cancel()
-
-	resp, err := hn.rpc.WalletKit.BumpFee(ctxt, req)
-	require.NoError(h, err, "unable to bump fee")
-
-	return resp
-}
-
-// PendingSweeps makes a RPC call to the node's WalletKitClient and asserts.
-func (h *HarnessTest) PendingSweeps(
-	hn *HarnessNode) *walletrpc.PendingSweepsResponse {
-
-	ctxt, cancel := context.WithTimeout(h.runCtx, DefaultTimeout)
-	defer cancel()
-
-	req := &walletrpc.PendingSweepsRequest{}
-
-	resp, err := hn.rpc.WalletKit.PendingSweeps(ctxt, req)
-	require.NoError(h, err, "unable to get pending sweeps")
-
-	return resp
-}
-
-// BuildRoute makes a RPC call to the node's RouterClient and asserts.
-func (h *HarnessTest) BuildRoute(hn *HarnessNode,
-	req *routerrpc.BuildRouteRequest) *routerrpc.BuildRouteResponse {
-
-	ctxt, cancel := context.WithTimeout(h.runCtx, DefaultTimeout)
-	defer cancel()
-
-	resp, err := hn.rpc.Router.BuildRoute(ctxt, req)
-	require.NoError(h, err, "failed to build route")
-	return resp
-}
-
-type InterceptorClient routerrpc.Router_HtlcInterceptorClient
-
-// HtlcInterceptor makes a RPC call to the node's RouterClient and asserts.
-func (h *HarnessTest) HtlcInterceptor(hn *HarnessNode) (
-	InterceptorClient, context.CancelFunc) {
-
-	// HtlcInterceptor needs to have the context alive for the entire test
-	// case as the returned client will be used for send and receive events
-	// stream. Thus we use cancel context here instead of a timeout
-	// context.
-	ctxt, cancel := context.WithCancel(h.runCtx)
-	resp, err := hn.rpc.Router.HtlcInterceptor(ctxt)
-
-	require.NoError(h, err, "failed to create HtlcInterceptor")
-
-	return resp, cancel
-}
-
-// ReceiveHtlcInterceptor waits until a message is received on the htlc
-// interceptor stream or the timeout is reached.
-func (h *HarnessTest) ReceiveHtlcInterceptor(
-	stream InterceptorClient) *routerrpc.ForwardHtlcInterceptRequest {
-
-	chanMsg := make(chan *routerrpc.ForwardHtlcInterceptRequest)
-	errChan := make(chan error)
-	go func() {
-		// Consume one message. This will block until the message is
-		// received.
-		resp, err := stream.Recv()
-		if err != nil {
-			errChan <- err
-			return
-		}
-		chanMsg <- resp
-	}()
-
-	select {
-	case <-time.After(DefaultTimeout):
-		require.Fail(h, "timeout", "timeout intercepting htlc")
-
-	case err := <-errChan:
-		require.Failf(h, "err from stream",
-			"received err from stream: %v", err)
-
-	case updateMsg := <-chanMsg:
-		return updateMsg
-	}
-
-	return nil
-}
-
-// SendOutputs makes a RPC call to the node's WalletKitClient and asserts.
-func (h *HarnessTest) SendOutputs(hn *HarnessNode,
-	req *walletrpc.SendOutputsRequest) *walletrpc.SendOutputsResponse {
-
-	ctxt, cancel := context.WithTimeout(h.runCtx, DefaultTimeout)
-	defer cancel()
-
-	resp, err := hn.rpc.WalletKit.SendOutputs(ctxt, req)
-	require.NoError(h, err, "failed to SendOutputs")
-
-	return resp
-}
-
-// SignPsbt makes a RPC call to the node's WalletKitClient and asserts.
-func (h *HarnessTest) SignPsbt(hn *HarnessNode,
-	req *walletrpc.SignPsbtRequest) *walletrpc.SignPsbtResponse {
-
-	ctxt, cancel := context.WithTimeout(h.runCtx, DefaultTimeout)
-	defer cancel()
-
-	resp, err := hn.rpc.WalletKit.SignPsbt(ctxt, req)
-	require.NoError(h, err, "failed to SignPsbt")
-
-	return resp
-}
-
-// ListAccounts makes a RPC call to the node's WalletKitClient and asserts.
-func (h *HarnessTest) ListAccounts(hn *HarnessNode,
-	req *walletrpc.ListAccountsRequest) *walletrpc.ListAccountsResponse {
-
-	ctxt, cancel := context.WithTimeout(h.runCtx, DefaultTimeout)
-	defer cancel()
-
-	resp, err := hn.rpc.WalletKit.ListAccounts(ctxt, req)
-	require.NoError(h, err, "failed to ListAccounts")
-
-	return resp
-}
-
-// ImportAccount makes a RPC call to the node's WalletKitClient and asserts.
-func (h *HarnessTest) ImportAccount(hn *HarnessNode,
-	req *walletrpc.ImportAccountRequest) *walletrpc.ImportAccountResponse {
-
-	ctxt, cancel := context.WithTimeout(h.runCtx, DefaultTimeout)
-	defer cancel()
-
-	resp, err := hn.rpc.WalletKit.ImportAccount(ctxt, req)
-	require.NoError(h, err, "failed to import account")
-
-	return resp
-}
-
-// PublishTransaction makes a RPC call to the node's WalletKitClient and
-// asserts.
-func (h *HarnessTest) PublishTransaction(hn *HarnessNode,
-	req *walletrpc.Transaction) *walletrpc.PublishResponse {
-
-	ctxt, cancel := context.WithTimeout(h.runCtx, DefaultTimeout)
-	defer cancel()
-
-	resp, err := hn.rpc.WalletKit.PublishTransaction(ctxt, req)
-	require.NoError(h, err, "failed to publish tx")
-
-	return resp
-}
-
-// ImportPublicKey makes a RPC call to the node's WalletKitClient and asserts.
-func (h *HarnessTest) ImportPublicKey(hn *HarnessNode,
-	req *walletrpc.ImportPublicKeyRequest) *walletrpc.ImportPublicKeyResponse {
-
-	ctxt, cancel := context.WithTimeout(h.runCtx, DefaultTimeout)
-	defer cancel()
-
-	resp, err := hn.rpc.WalletKit.ImportPublicKey(ctxt, req)
-	require.NoError(h, err, "failed to import public key")
-
-	return resp
-}
-
-type SingleInvoiceClient invoicesrpc.Invoices_SubscribeSingleInvoiceClient
-
-// SubscribeSingleInvoice creates a subscription client for given invoice and
-// asserts its creation.
-func (h *HarnessTest) SubscribeSingleInvoice(hn *HarnessNode,
-	rHash []byte) SingleInvoiceClient {
-
-	req := &invoicesrpc.SubscribeSingleInvoiceRequest{RHash: rHash}
-
-	// SubscribeSingleInvoice needs to have the context alive for the
-	// entire test case as the returned client will be used for send and
-	// receive events stream. Thus we use runCtx here instead of a timeout
-	// context.
-	client, err := hn.rpc.Invoice.SubscribeSingleInvoice(h.runCtx, req)
-	require.NoError(h, err, "unable to create single invoice client")
-
-	return client
-}
-
-// ReceiveSingleInvoice waits until a message is received on the subscribe
-// single invoice stream or the timeout is reached.
-func (h *HarnessTest) ReceiveSingleInvoice(
-	stream SingleInvoiceClient) *lnrpc.Invoice {
-
-	chanMsg := make(chan *lnrpc.Invoice)
-	errChan := make(chan error)
-	go func() {
-		// Consume one message. This will block until the message is
-		// received.
-		resp, err := stream.Recv()
-		if err != nil {
-			errChan <- err
-			return
-		}
-		chanMsg <- resp
-	}()
-
-	select {
-	case <-time.After(DefaultTimeout):
-		require.Fail(h, "timeout", "timeout receiving single invoice")
-
-	case err := <-errChan:
-		require.Failf(h, "err from stream",
-			"received err from stream: %v", err)
-
-	case updateMsg := <-chanMsg:
-		return updateMsg
-	}
-
-	return nil
-}
-
-type MiddlewareClient lnrpc.Lightning_RegisterRPCMiddlewareClient
-
-// RegisterRPCMiddleware makes a RPC call to the node's RegisterRPCMiddleware
-// and asserts. It also returns a cancel context which can cancel the context
-// used by the client.
-func (h *HarnessTest) RegisterRPCMiddleware(
-	hn *HarnessNode) (MiddlewareClient, context.CancelFunc) {
-
-	ctxt, cancel := context.WithCancel(h.runCtx)
-
-	stream, err := hn.rpc.LN.RegisterRPCMiddleware(ctxt)
-	require.NoError(h, err, "failed to register rpc middleware")
-
-	return stream, cancel
 }
 
 // MineEmptyBlocks mines a given number of empty blocks.
@@ -3921,106 +1151,4 @@ func (h *HarnessTest) MineBlockWithTxes(txes []*btcutil.Tx) *wire.MsgBlock {
 	require.NoError(h, err, "unable to get block")
 
 	return block
-}
-
-// AssertTransactionInWallet asserts a given txid can be found in the node's
-// wallet.
-func (h *HarnessTest) AssertTransactionInWallet(hn *HarnessNode,
-	txid chainhash.Hash) {
-
-	err := wait.NoError(func() error {
-		txResp := h.GetTransactions(hn)
-		for _, txn := range txResp.Transactions {
-			if txn.TxHash == txid.String() {
-				return nil
-			}
-		}
-
-		return fmt.Errorf("%s: expected txid=%v not found in wallet",
-			hn.Name(), txid)
-	}, DefaultTimeout)
-
-	require.NoError(h, err, "failed to find tx")
-}
-
-// AssertTransactionNotInWallet asserts a given txid can NOT be found in the
-// node's wallet.
-func (h *HarnessTest) AssertTransactionNotInWallet(hn *HarnessNode,
-	txid chainhash.Hash) {
-
-	err := wait.NoError(func() error {
-		txResp := h.GetTransactions(hn)
-		for _, txn := range txResp.Transactions {
-			if txn.TxHash == txid.String() {
-				return fmt.Errorf("expected txid=%v to be "+
-					"not found", txid)
-			}
-		}
-
-		return nil
-	}, DefaultTimeout)
-
-	require.NoError(h, err, "failed to assert tx not found")
-}
-
-// GetOutputIndex returns the output index of the given address in the given
-// transaction.
-func (h *HarnessTest) GetOutputIndex(txid *chainhash.Hash, addr string) int {
-	h.Helper()
-
-	// We'll then extract the raw transaction from the mempool in order to
-	// determine the index of the p2tr output.
-	tx := h.GetRawTransaction(txid)
-
-	p2trOutputIndex := -1
-	for i, txOut := range tx.MsgTx().TxOut {
-		_, addrs, _, err := txscript.ExtractPkScriptAddrs(
-			txOut.PkScript, h.Miner().ActiveNet,
-		)
-		require.NoError(h, err)
-
-		if addrs[0].String() == addr {
-			p2trOutputIndex = i
-		}
-	}
-	require.Greater(h, p2trOutputIndex, -1)
-
-	return p2trOutputIndex
-}
-
-type ConfNtfnClient chainrpc.ChainNotifier_RegisterConfirmationsNtfnClient
-
-// RegisterConfirmationsNtfn creates a notification client to watch a given
-// transaction being confirmed.
-func (h *HarnessTest) RegisterConfirmationsNtfn(hn *HarnessNode,
-	req *chainrpc.ConfRequest) ConfNtfnClient {
-
-	// RegisterConfirmationsNtfn needs to have the context alive for the
-	// entire test case as the returned client will be used for send and
-	// receive events stream. Thus we use runCtx here instead of a timeout
-	// context.
-	client, err := hn.rpc.ChainClient.RegisterConfirmationsNtfn(
-		h.runCtx, req,
-	)
-	require.NoError(h, err, "unable to register conf client")
-
-	return client
-}
-
-type SpendClient chainrpc.ChainNotifier_RegisterSpendNtfnClient
-
-// RegisterSpendNtfn creates a notification client to watch a given
-// transaction being spent.
-func (h *HarnessTest) RegisterSpendNtfn(hn *HarnessNode,
-	req *chainrpc.SpendRequest) SpendClient {
-
-	// RegisterSpendNtfn needs to have the context alive for the entire
-	// test case as the returned client will be used for send and receive
-	// events stream. Thus we use runCtx here instead of a timeout context.
-	client, err := hn.rpc.ChainClient.RegisterSpendNtfn(
-		h.runCtx, req,
-	)
-	require.NoError(h, err, "unable to register spend client")
-
-	return client
 }
